Algorithm.BinarySearch

Description

\(O(log N)\) binary search for sorted items in an inclusive range (from left to right only).

bsearch returns an (ok, ng) index pair at the boundary. bsearchL and bsearchR returns one of the pair. bsearchM is a monadic variant of bsearch.

Example

With an OK predicate (<= 5), list [0..9] can be seen as:

[0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
 <-------------->  <-------->
        ok             ng

In the preceding example, bsearch returns the (ok, ng) = (Just 5, Just 6) pair at the boundary:

>>> :{
let xs = [0 :: Int .. 9]
 in bsearch (0 :: Int, 9 :: Int) (\i -> xs !! i <= 5)
:}
(Just 5,Just 6)

bsearchL returns Just 5 and bsearchR returns Just 6.

Documentation

bsearchLG :: Vector v a => v a -> (a -> Bool) -> Maybe a

bsearchL over a vector.

bsearchExact :: (Vector v a, Ord b) => v a -> (a -> b) -> b -> Maybe a

bsearchL over a vector, searching for a specific value. FIXME: It's slower than bsearchLG.

bsearch :: (Int, Int) -> (Int -> Bool) -> (Maybe Int, Maybe Int)

Pure binary search.

bsearchL :: (Int, Int) -> (Int -> Bool) -> Maybe Int

Also known as lower bound.

bsearchR :: (Int, Int) -> (Int -> Bool) -> Maybe Int

Also known as upper bound.

bsearchM :: forall m. Monad m => (Int, Int) -> (Int -> m Bool) -> m (Maybe Int, Maybe Int)

Monadic binary search.

bsearchML :: forall m. Monad m => (Int, Int) -> (Int -> m Bool) -> m (Maybe Int)

bsearchMR :: forall m. Monad m => (Int, Int) -> (Int -> m Bool) -> m (Maybe Int)

bsearchF32 :: (Float, Float) -> Float -> (Float -> Bool) -> (Maybe Float, Maybe Float)

bsearchF32L :: (Float, Float) -> Float -> (Float -> Bool) -> Maybe Float

bsearchF32R :: (Float, Float) -> Float -> (Float -> Bool) -> Maybe Float

bsearchF64 :: (Double, Double) -> Double -> (Double -> Bool) -> (Maybe Double, Maybe Double)

bsearchF64L :: (Double, Double) -> Double -> (Double -> Bool) -> Maybe Double

bsearchF64R :: (Double, Double) -> Double -> (Double -> Bool) -> Maybe Double

compressList :: [Int] -> (Vector Int, [Int])

One dimensional index compression: xs -> (nubSorted, indices)

isqrtSlow :: Int -> Int

Retrieves square root of an Int.