Data.AdhocGraph

Description

Graph search with context for every vertex in the super graph.

Tree search doesn't need contect per vertex and thus not the target. But it would be great if we can use he same function for both trees and graphs.

Common cases

Search

[ ] DFS

[ ] BFS

[ ] Dijkstra

Search tree

[ ] Search paths

Search tree

[ ] DFS tree

[ ] BFS tree

[ ] Dijkstra tree

[ ] 01-BFS

Niche cases

[ ] Tree diameter

[ ] Walk through every path C - Remembering the Days

[ ] Come back to the first vertex

Documentation

data AdhocGraph m w

Adhoc graph with distance-based search context.

Constructors

AdhocGraph

Fields

undefAG :: !w

Initial weights for non-visited vertices.

nVertsAG :: !Int

The number of vertices.

visAG :: !(MVector (PrimState m) Bool)

Visibility context.

distAG :: !(MVector (PrimState m) w)

Distance context.

parentAG :: !(MVector (PrimState m) Int)

Distance context.

adjAG :: !(Vertex -> Vector Vertex)

Non-weightened graph representation as a function.

adjWAG :: !(Vertex -> Vector (Vertex, w))

Weightened graph representation as a function.

data AdhocMethod m w

Adhoc graph search methods for weightened graphs.

Non-weighnted graphs need to be represented as a graph with weight `1`.

Constructors

AdhocMethod

Fields

markRootAM :: !(AdhocGraph m w -> Vertex -> w -> m ())

Marks a root vertex.

markAM :: !(AdhocGraph m w -> Vertex -> (Vertex, w) -> m ())

Marks a child vertex.

unmarkAM :: !(AdhocGraph m w -> Vertex -> (Vertex, w) -> m ())

Unmarks a child vertex.

wfilterAM :: !(AdhocGraph m w -> Vertex -> (Vertex, w) -> m Bool)

Filter for adjacent vertices on DFS.

defaultMarkRootAM :: (PrimMonad m, Unbox w) => AdhocGraph m w -> Vertex -> w -> m ()

defaultMarkAM :: (PrimMonad m, Unbox w, Eq w, Num w) => AdhocGraph m w -> Vertex -> (Vertex, w) -> m ()

defaultUnmarkAM :: (PrimMonad m, Unbox w) => AdhocGraph m w -> Vertex -> (Vertex, w) -> m ()

defaultWFilterAM :: (PrimMonad m, Unbox w, Eq w) => AdhocGraph m w -> Vertex -> (Vertex, w) -> m Bool

amDefault :: (PrimMonad m, Unbox w, Num w, Eq w) => AdhocMethod m w

amDefaultInt :: PrimMonad m => AdhocMethod m Int

data AdhocGraphArgs w

Constructors

AdhocGraphArgs

Fields

undefAGA :: !w

Initial weights for non-visited vertices.

nVertsAGA :: !Int

The number of vertices.

distAGA :: !Bool

Distance context.

visAGA :: !Bool

Distance context.

parentAGA :: !Bool

The capacity of the parents.

adjAGA :: !(Vertex -> Vector Vertex)

Non-weightened graph representation as a function.

adjWAGA :: !(Vertex -> Vector (Vertex, w))

Weightened graph representation as a function.

agaDefaultUnit :: AdhocGraphArgs ()

agaDefaultInt :: AdhocGraphArgs Int

newAg :: (PrimMonad m, Unbox w) => AdhocGraphArgs w -> m (AdhocGraph m w)

adjWFilterAM :: (PrimMonad m, Unbox w) => AdhocMethod m w -> AdhocGraph m w -> Vertex -> m (Vector (Vertex, w))

markScopeAM :: PrimMonad m => AdhocMethod m w -> AdhocGraph m w -> Vertex -> (Vertex, w) -> m () -> m ()

Marks the next vertex while the closure is running.

distFilterAG :: (PrimMonad m, Unbox w, Ord w) => AdhocGraph m w -> Vector Vertex -> m (Vector Vertex)

Filter by distAG: equals to undefAG or not.

distFilterWAG :: (PrimMonad m, Unbox w, Ord w) => AdhocGraph m w -> Vector (Vertex, w) -> m (Vector (Vertex, w))

Filter by distAG: equals to undefAG or not.

visFilterAG :: PrimMonad m => AdhocGraph m w -> Vector Vertex -> m (Vector Vertex)

Filter by visAG

markDistAG :: (PrimMonad m, Unbox w, Num w) => AdhocGraph m w -> Vertex -> (Vertex, w) -> m ()

Marks distance.

markVisAG :: (PrimMonad m, Unbox w, Num w) => AdhocGraph m w -> Vertex -> (Vertex, w) -> m ()

Marks visibility.

unmarkDistAG :: (PrimMonad m, Unbox w) => AdhocGraph m w -> Vertex -> (Vertex, w) -> m ()

Unmarks distance.

unmarkVisAG :: PrimMonad m => AdhocGraph m w -> Vertex -> Vertex -> m ()

Unmarks visibility.

markParentAG :: PrimMonad m => AdhocGraph m w -> Vertex -> (Vertex, w) -> m ()

distAdjWAG :: (PrimMonad m, Unbox w, Ord w) => AdhocGraph m w -> Vertex -> m (Vector (Vertex, w))

For DFS and BFS.

dfsAG :: (PrimMonad m, Unbox w, Num w, Ord w) => AdhocGraph m w -> AdhocMethod m w -> Vertex -> m ()

O(V+E) Runs depth-first search.

dfsAllAG :: (PrimMonad m, Unbox w, Num w) => AdhocGraph m w -> AdhocMethod m w -> Vertex -> m ()

O(N!) Runs depth-first search in every path. REMARK: Consider using tspDP instead.

dfsBackAllAG :: (PrimMonad m, Unbox w) => AdhocGraph m w -> AdhocMethod m w -> Vertex -> m ()

O(N!) Runs depth-first search in every path, returning to the start. REMARK: Consider using tspDP instead.

bfsAG :: (PrimMonad m, Unbox w, Num w, Ord w) => AdhocGraph m w -> AdhocMethod m w -> Vertex -> m ()

O(V+E) Runs breadth-first search.

pruneAdjWAG :: (PrimMonad m, Unbox w, Ord w, Num w) => AdhocGraph m w -> Vertex -> w -> m (Vector (Vertex, w))

For 01-BFS or Dijkstra.

bfs01AG :: (PrimMonad m, Unbox w, Ord w, Num w) => AdhocGraph m w -> AdhocMethod m w -> Vertex -> m ()

01-BFS. TODO: Super vertices?

djAG :: (PrimMonad m, Unbox w, Ord w, Num w) => AdhocGraph m w -> AdhocMethod m w -> Vertex -> m ()

O(V+E) Runs breadth-first search.