Data.Core.SegmentTreeAction

Description

Right semigroup action, specialized for the lazily propagated segment tree.

Algebra

\(s_2 * (s_1 * a) == (s_2 \diamond s_1) * a\)

Documentation

class SegmentTreeAction op a where

Right semigroup action, specialized for the lazily propagated segment tree.

Minimal complete definition

Nothing

Methods

segAct :: op -> a -> a

Right semigroup aciton, limited to the lazily propagated segment tree.

segActWithLength :: op -> a -> Int -> a

Right semigroup aciton with the segment length.