Data.Graph

Description

I'm using Array Int [Int] as a primary Graph data storage.

Documentation

type Graph a = Array Int [a]

Adjacency list representation of a graph with node type parameter a.

type Vertex = Int

Vertex index. Starts with zero.

type WGraph a = Array Int [Entry a Vertex]

Weighted Graph. Entry priority payload.

bfsVec :: Graph Int -> Int -> Vector Int

Collects distances from one vertex to every other using BFS, returning a vector.

bfsPath :: Graph Int -> Int -> Int -> Maybe Int

BFS template for finding a shortest path from one vertex to another.

bfsVerts :: Graph Int -> Int -> IntMap Int

BFS template for collecting shortest paths from one vertex to every other.

bfsGrid :: UArray (Int, Int) Char -> (Int, Int) -> UArray (Int, Int) Int

BFS over grid. Not generalized (yet).

solve01BFS :: (Int, Int) -> UArray (Int, Int) Char -> UArray (Int, Int) Int

Easy 01-BFS

Typical problems

ABC 176 D

bfsGrid01 :: (Int, Int) -> UArray (Int, Int) Bool -> UArray (Int, Int, Int) Int

Direction-based 01-BFS: https://atcoder.jp/contests/typical90/tasks/typical90_aq. It's slow, but could be applied easily in certain situations.

Typical problems

Typoical 043 - Maze Challenge with Lack of Sleep (★4)

components :: Graph Int -> Int -> IntSet

DFS where all the reachable vertices from one vertex are collcetd.

dfsEveryPathT072 :: UArray (Int, Int) Char -> (Int, Int) -> Int

DFS for every path, specially for Typical 072.

cyclesSUG :: Array Vertex [Vertex] -> Vector Bool

Checks a Simple Undirected Graph and returns markings of cycle vertices. TODO: Test if it works as expected.

dj :: forall a. (Num a, Ord a) => WGraph a -> Int -> IntMap a

Dijkstra template that collects all the shortest distances from one vertex to every other. Works for weightened graphs with positive edge capacities only.

Pro tip: Use reverse graph to collect cost from every other vertex to one (see revDjAll).

revDj :: WGraph Int -> Int -> IntMap Int

Runs Dijkstra's algorithm over a reversed graph of given graph.

revWGraph :: WGraph Int -> WGraph Int

Creates a reverse weightened graph.

djVec :: forall a. (Num a, Ord a, Unbox a) => WGraph a -> Int -> a -> Vector a

Dijkstra backed by a vector.

revDjVec :: WGraph Int -> Int -> Vector Int

Runs Dijkstra's algorithm over a reversed graph of given graph.