Data.Graph.Dinic

Description

O(V^2E) max flow algorithm (Dinic's algorithm). Heavily inspired by cojna/iota.

Just run maxFlowD to get the result. All the other functions are used internally.

Documentation

data Dinic s c

The state for Dinic's algorithm

Internals

Internally the graph is stored in a CSR (compressed sparse row).

Constructors

Dinic

Fields

nVertsD :: !Int

nEdgesD :: !Int

offsetsD :: !(Vector Int)

Source vertex -> initial edge index. Note that edges are sourted by the starting vertex.

edgeDstD :: !(Vector Int)

Edge index -> destination vertex.

edgeRevIndexD :: !(Vector Int)

Edge index -> reverse edge index.

edgeCapD :: !(MVector s c)

Edge index -> residual edge capacity.

data DinicBuffer s c

Mutable states on running Dinic's algorithm.

Constructors

DinicBuffer

Fields

distsD :: !(MVector s Int)

BFS result.

queueD :: !(Buffer s Vertex)

Queue for BFS

iterD :: !(MVector s Int)

DFS state for iterating through all the neighbors.

maxFlowD :: (Unbox c, Num c, Ord c, Bounded c) => Int -> Int -> Int -> Vector (Vertex, Vertex, c) -> c

O(V^2E) max flow algorithm (Dinic's algorithm)

undefD :: Int

buildDinic :: forall c m. (Unbox c, Num c, PrimMonad m) => Int -> Vector (Vertex, Vertex, c) -> m (Dinic (PrimState m) c)

Builds Dinic from edges.

runDinic :: forall c m. (Unbox c, Num c, Ord c, Bounded c, PrimMonad m) => Vertex -> Vertex -> Dinic (PrimState m) c -> m c

Runs the Dinic's algorithm.

Overview

Dinic's algorithm loops the following steps: 1. runDinicBfs: Creates a DAG with BFS. 2. runDinicDfs: Adds flow from the source to the sink.

runDinicBfs :: forall c m. (Unbox c, Num c, Ord c, PrimMonad m) => Vertex -> Vertex -> Dinic (PrimState m) c -> DinicBuffer (PrimState m) c -> m ()

Collect shortest distances from the source vertex using BFS.

runDinicDfs :: forall c m. (Unbox c, Num c, Ord c, PrimMonad m) => Vertex -> Vertex -> c -> Dinic (PrimState m) c -> DinicBuffer (PrimState m) c -> m c

Modify the flow