Data.Graph.MaxFlow

Description

Maximum flow calculation (Ford-Fulkerson algorithm).

Documentation

data RNEdge

Edge in residual network from on vertex to another.

Constructors

RNEdge

Fields

to :: !Vertex

Points the the other side of the edge

cap :: !Int

Capacity of the edge, or the flow from the vertex to another

rev :: !Int

The other side of the vertices is pointed with `rn ! (rev (rn ! to))` so that edge insertion takes just $O(1)$.

type ResidualNetwork = IOVector (IntMap RNEdge)

Vertex -> [RNEdge].

TODO: For the sub containers, use Sequence or something better

buildRN :: Int -> [(Int, (Int, Int))] -> IO ResidualNetwork

Builds a residual network at initial state from given edges `(v2, cost)`. {-# INLINE buildRN #-} TODO: make it generic over ST.. for no reason?

maxFlowRN :: Int -> ResidualNetwork -> Int -> Int -> IO Int

Calculates max flow.

augumentPath :: ResidualNetwork -> IOVector Bool -> Vertex -> Int -> IO (Maybe (Int, [(Vertex, Vertex)]))

Find a flow augment path between two vertices.

updateFlow :: ResidualNetwork -> Int -> [(Vertex, Vertex)] -> IO ()

addFlowRNEdge :: ResidualNetwork -> Vertex -> Vertex -> Int -> IO ()