Data.Graph.MinCostFlow

Description

Minimum cost flow calculation (Ford-Fulkerson algorithm).

Typical problems

min_cost_b_flow

Documentation

type CapacityMCF c = c

type FlowMCF c = c

type CostMCF c = c

data MinCostFlow s c

The state for CostFlow

Internals

Internally the graph is stored in a CSR (compressed sparse row).

Constructors

MinCostFlow

Fields

nVertsMCF :: !Int

nEdgesMCF :: !Int

offsetsMCF :: !(Vector Int)

Source vertex -> initial edge index. Note that edges are sorted by the starting vertex.

edgeDstMCF :: !(Vector Int)

Edge index -> destination vertex.

edgeRevIndexMCF :: !(Vector Int)

Edge index -> reverse edge index.

edgeCapMCF :: !(MVector s (CapacityMCF c))

Edge index -> residual edge capacity.

edgeCostMCF :: !(Vector (CostMCF c))

Edge index -> cost

data MinCostFlowBuffer r s c

Mutable states on running MinCostFlow algorithm.

Constructors

MinCostFlowBuffer

Fields

distsMCF :: !(MVector s r)

Vertex -> shortest wrapped cost (Bellman-Ford result)

prevVertMCF :: !(MVector s Vertex)

Vertex -> shortest distance + last vertex

prevEdgeMCF :: !(MVector s EdgeId)

Vertex -> shortest distance + last edge

relaxedCostFlow' :: (Show (f (CostMCF c)), Num (f (CostMCF c)), Monoid (f (CostMCF c)), Unbox (f (CostMCF c)), Ord (f (CostMCF c)), Num (f (CostMCF c)), PrimMonad m, Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => (CostMCF c -> f (CostMCF c)) -> Int -> Int -> Int -> CapacityMCF c -> Vector (Vertex, Vertex, CapacityMCF c, CostMCF c) -> m (Maybe (f (CostMCF c)), MinCostFlow (PrimState m) c)

Returns the wrapped cost for getting the flow, or Nothing when impossible.

TODO: Return tuple just like ACL.

relaxedCostFlow :: (Show (f (CostMCF c)), Num (f (CostMCF c)), Monoid (f (CostMCF c)), Unbox (f (CostMCF c)), Ord (f (CostMCF c)), Num (f (CostMCF c)), Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => (CostMCF c -> f (CostMCF c)) -> Int -> Int -> Int -> CapacityMCF c -> Vector (Vertex, Vertex, CapacityMCF c, CostMCF c) -> Maybe (f (CostMCF c))

minCostFlow :: (Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => Int -> Int -> Int -> CapacityMCF c -> Vector (Vertex, Vertex, CapacityMCF c, CostMCF c) -> Maybe (Min (CostMCF c))

Returns the minimum cost for getting the flow, or Nothing when impossible.

TODO: Return tuple just like ACL.

minCostFlow' :: (PrimMonad m, Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => Int -> Int -> Int -> CapacityMCF c -> Vector (Vertex, Vertex, CapacityMCF c, CostMCF c) -> m (Maybe (Min (CostMCF c)), MinCostFlow (PrimState m) c)

maxCostFlow :: (Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => Int -> Int -> Int -> CapacityMCF c -> Vector (Vertex, Vertex, CapacityMCF c, CostMCF c) -> Maybe (Max (CostMCF c))

Returns the minimum cost for getting the flow, or Nothing when impossible.

TODO: Return tuple just like ACL.

maxCostFlow' :: (PrimMonad m, Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => Int -> Int -> Int -> CapacityMCF c -> Vector (Vertex, Vertex, CapacityMCF c, CostMCF c) -> m (Maybe (Max (CostMCF c)), MinCostFlow (PrimState m) c)

edgesMCF :: (PrimMonad m, Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => MinCostFlow (PrimState m) c -> m (Vector (Int, Int, CapacityMCF c, FlowMCF c, CostMCF c))

Retrieves edge information (v1, v2, cap, flow, cost) from the maxFlow results.

Be warned that it contains reverse edges and edges fromto sourcesink.

undefMCF :: Int

For vertices and edge indices?

buildMinCostFlow :: forall c m. (PrimMonad m, Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => Int -> Vector (Vertex, Vertex, CapacityMCF c, CostMCF c) -> m (MinCostFlow (PrimState m) c)

Builds MinCostFlow from edges.

runMinCostFlow :: forall f c m. (Show (f (CostMCF c)), Num (f (CostMCF c)), Monoid (f (CostMCF c)), Unbox (f (CostMCF c)), Ord (f (CostMCF c)), Num (f (CostMCF c)), PrimMonad m, Show c, Num c, Unbox c, Integral c, Ord c, Bounded c) => (CostMCF c -> f (CostMCF c)) -> Vertex -> Vertex -> FlowMCF c -> MinCostFlow (PrimState m) c -> m (Maybe (f (CostMCF c)))

Runs the MinCostFlow algorithm.

runMinCostFlowShortests :: (Show (f (CostMCF c)), Unbox (f (CostMCF c)), Monoid (f (CostMCF c)), Ord (f (CostMCF c)), Num (f (CostMCF c)), Show c, Unbox c, Num c, Ord c, Bounded c, PrimMonad m) => (CostMCF c -> f (CostMCF c)) -> Vertex -> MinCostFlow (PrimState m) c -> MinCostFlowBuffer (f (CostMCF c)) (PrimState m) c -> m ()

Collect shortest distances from the source vertex using Bellman-Ford algorithm.

Hangs on negative loop.

TODO: break? TODO: replace with dijkstra with potencials.