Data.Graph.Tree.Hld

Description

Heavy-light decomposition. Heavily inspired by cojna/iota.

Documentation

type VertexHLD = Vertex

Vertex re-ordered by HLD.

data HLD

Heavy-light decomposition.

Constructors

HLD

Fields

parentHLD :: Vector Vertex

Vertex -> Parent vertex.

indexHLD :: Vector VertexHLD

Vertex -> Reorderd index.

pathHeadHLD :: Vector Vertex

Vertex -> Their path's head vertex.

lcaHLD :: HLD -> Vertex -> Vertex -> Vertex

\(O(log V)\).

Typical Problems

ABC 133 - F

pathHLD :: HLD -> Vertex -> Vertex -> Vector (VertexHLD, VertexHLD)

\(O(log V)\) Returns inclusive vertex pairs per HLD path. - TODO: know the details - TODO: consider direction

foldCommuteHLD :: (Monoid mono, Monad m) => HLD -> (VertexHLD -> VertexHLD -> m mono) -> Vertex -> Vertex -> m mono

Folds commutative monoid on a tree using HLD.

Segment tree

HLD path folding is often done with a segment tree. It uses VertexHLD as indices. If edges have weights, you can either treat edges as new vertices or put weight to the depper index (when (v1 /= v2) always holds.

Idea 1. Edges as new vertices

o--o--o  -> o-x-o-x-o

Idea 2. Put weight to deeper vertex

  o
  | <--- edge 1
  o <- write w1 here
  | <--- edge 1
  o <- write w2 here

Typical Problems

ABC 294 - G

hldOf :: forall w. SparseGraph Int w -> HLD

Heavy-light decomposition or Centroid Path Decomposition.

About

HLD builds a smaller tree on top of an existing tree, combining vertices into paths.

References

https://take44444.github.io/Algorithm-Book/graph/tree/hld/main.html