Data.Graph.Tree.TreeSG

Description

Tree algorithms for SparseGraph.

Documentation

treeDiameterSG :: SparseGraph Int w -> Int

\(O(V+E)\) Tree diameter calculation.

treeDepthInfoSG :: (Monoid w, Unbox w) => SparseGraph Int w -> Int -> (Vector Int, TransitionalSemigroup w)

\(O(N)\) Returns (depths, parents).

lcaCacheSG :: (Monoid w, Unbox w) => SparseGraph Int w -> Vertex -> LcaCache w

Returns LcaCache, i.e., (parents, depths, parents').

foldTreeImpl :: forall m op a w. Monad m => SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> (Vertex -> a -> m ()) -> m a

foldTree :: SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> a

\(O(N)\) Folds a tree from one root vertex using postorder DFS.

scanTree :: Vector v a => SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> v a

\(O(N)\) Folds a tree from one root vertex using postorder DFS, recording all the accumulation values on every vertex.

scanTreeU :: Unbox a => SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> Vector a

\(O(N)\) Type-restricted scanTree.

scanTreeV :: SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> Vector a

\(O(N)\) Type-restricted scanTree.

foldTreeAllSG :: forall a op w. (Unbox a, Unbox op, Monoid op, SemigroupAction op a) => SparseGraph Int w -> (Vertex -> a) -> (a -> op) -> Vector a

\(O(N)\) Folds a tree for every vertex as a root using the rerooting technique. REMARK: mempty is used for initial operator value.

Typical problems

Typical 039 - Tree Distance (★5)

EDPC V - Subtree

TDPC N - tree