Data.Instances.A2

Description

Unboxed 2D array stored in a packed 1-dimensional vector.

The Unbox implementation is from Affine type in cojna/iota.

Documentation

data A2 a

Unboxed 2D array packed in one array when stored in Vector.

Constructors

A2 !a !a