Data.Instances.A3

Description

Unboxed 3D array stored in a packed 1-dimensional vector.

The Unbox implementation is from Affine type in cojna/iota.

Documentation

data A3 a

Unboxed 3D array packed in one array when stored in Vector.

Constructors

A3 !a !a !a