Data.RollingHash

Description

The rolling hash algorithm lets you create fastly ((O(1))) comparable / concatanatable string slice in after \(O(N)\) preparation.

I suspect if slices longer than the orignal string can be calculated without panic in my implementation.

Documentation

data RollingHash b p

Rolling hash of a string.

Internals

Slice (2, 4) of "abcdef" is given as this:

           s :=     a       b       c       d       e
           s4 = b^4 a + b^3 b + b^2 c + b^1 d + b^0 e
           s2 = b^1 a + b^0 b
s4 - s2 * b^3 =                 b^2 c + b^1 d + b^0 e

Constructors

RollingHash

Fields

sourceLength :: !Int

dimensions :: !(Vector Int)

${B^i mod p}_{i elem [0, n)}$

hashSum :: !(Vector Int)

data HashInt

Type that represents a B-adic number for the rolling hash algorithm.

Constructors

HashInt

newRH :: forall p. TypeInt p => String -> RollingHash HashInt p

Creates a rolling hash of given string.

lengthRH :: RollingHash b p -> Int

Retrieves the original length of the RollingHash string.

data HashSlice p

HashSlice value length. See also the example of RollingHash.

Constructors

HashSlice

Fields

hashValue :: !Int

hashLength :: !Int

sliceRH :: forall b p. (TypeInt b, TypeInt p) => RollingHash b p -> Int -> Int -> HashSlice p

Slices a rolling hash string.

consHS :: forall b p. (TypeInt b, TypeInt p) => RollingHash b p -> HashSlice p -> HashSlice p -> HashSlice p

Cons two rolling hash slices.

emptyHS :: TypeInt p => HashSlice p

Creates an empty rolling hash slice.

concatHS :: forall b p t. (TypeInt b, TypeInt p, Foldable t) => RollingHash b p -> t (HashSlice p) -> HashSlice p

Concatanates two rolling hash slices.