Data.SegmentTree.Beats.SumMinMax

Description

SumMinMax and AddChminChmax on the segment tree beats.

Typical problems

Range Chmin Chmax Add Range Sum

Documentation

data SumMinMax a

Sum + Min + Max information on segment tree beats.

REMARK: Be sure to NOT use mempty as initial values in segment tree beats, use singletonSMM 0 instead.

TODO: Specialize and UNPACK.

Constructors

SumMinMax

Fields

sumSMM :: !a

minSMM :: !a

maxSMM :: !a

nMinSMM :: !Int

The number of elements that equals to minSMM in the interval.

nMaxSMM :: !Int

The number of elements that equals to maxSMM in the interval.

min2SMM :: !a

Secondary minimum value.

max2SMM :: !a

Secondary maximum value.

failsSMM :: !Bool

singletonSMM :: a -> SumMinMax a

data AddChminChmax a

Add first then apply chmin and chmax, works on segment tree beats.

TODO: Specialize and UNPACK.

Constructors

AddChminChmax

Fields

addACC :: !a

chminACC :: !a

chmaxACC :: !a

newAddACC :: (Num a, Ord a, Bounded a) => a -> AddChminChmax a

newChminACC :: (Num a, Ord a, Bounded a) => a -> AddChminChmax a

newChmaxACC :: (Num a, Ord a, Bounded a) => a -> AddChminChmax a

Unbox instances

type SumMinMaxRepr a = (a, a, a, a, a, (Int, Int, Bool))

type AddChminChmaxRepr a = (a, a, a)