Data.SparseGraph

Description

vector-based sparse graph implementation. Heavily inspired by cojna/iota.

Documentation

type Edge = (Vertex, Vertex)

type WEdgeWith w = (Vertex, Vertex, w)

Weightened edge

type EdgeId = Int

data SparseGraph i w

CSR (compressed sparse row) representation of a graph, weightened or unweightened.

Constructors

SparseGraph

Fields

boundsSG :: !(i, i)

Vertex index boundary.

nVertsSG :: !Int

Number of vertices.

nEdgesSG :: !Int

Number of edges.

offsetsSG :: !(Vector Int)

Maps Vector to the starting edge index.

adjacentsSG :: !(Vector Vertex)

Adjacent vertices sorted with starting vertex.

edgeWeightsSG :: !(Vector w)

Edge weight information.

buildUSG :: Unindex i => (i, i) -> Vector (i, i) -> SparseGraph i ()

Builds an unweightned SparseGraph.

buildWSG :: (Unindex i, Unbox w) => (i, i) -> Vector (i, i, w) -> SparseGraph i w

Builds a weightned SparseGraph.

buildRawSG :: (Unindex i, Unbox w) => (i, i) -> Vector (Vertex, Vertex, w) -> SparseGraph i w

adj :: SparseGraph i w -> Vertex -> Vector Vertex

Retrieves adjacent vertices.

eAdj :: SparseGraph i w -> Vertex -> Vector (EdgeId, Vertex)

Returns (EdgeId, Vertex) paris. Hardly used.

adjIx :: Unindex i => SparseGraph i w -> i -> Vector i

Retrieves adjacent vertex indices.

adjW :: Unbox w => SparseGraph i w -> Vertex -> Vector (Vertex, w)

Retrieves adjacent vertices with weights.

adjWIx :: (Unindex i, Unbox w) => SparseGraph i w -> i -> Vector (i, w)

Retrieves adjacent vertex indices with weights.

dfsSG :: Unindex i => SparseGraph i w -> i -> IxVector i (Vector Int)

O(V+E) Depth-first search. Returns a vector of distances to each vertex. Unreachable vertices are given distance of `-1`.

dfsEveryPathSG :: SparseGraph Int Int -> Int -> Int

O(V+E) Depth-first search. Just a template.

Typical problems

ABC 317 C - Remembering the Days

componentsVecSG :: Ix i => SparseGraph i w -> i -> IxVector i (Vector Bool)

O(V+E) Marks connected vertices. Also consider using union-find tree.

bfsSG :: Ix i => SparseGraph i w -> i -> IxVector i (Vector Int)

O(V+E) breadth-first search. Unreachable vertices are given distance of -1.

genericBfs :: (Int -> Vector Int) -> Int -> Vertex -> Vector Int

O(V+E) breadth-first search. Unreachable vertices are given distance of -1.

genericBfs01 :: (Ix i, Unbox i) => (i, i) -> (i -> Vector (i, Int)) -> Vector i -> IxUVector i Int

O(V+E) 01-BFS. Unreachable vertices are given distance of -1.

djSG :: forall i w. (Ix i, Unbox i, Num w, Ord w, Unbox w) => SparseGraph i w -> w -> Vector i -> IxUVector i w

O((E+V) log {V}) Dijkstra's algorithm.

genericDj :: forall w. (Unbox w, Num w, Ord w) => (Int -> Vector (Int, w)) -> Int -> w -> Vector Int -> Vector w

O((E+V) log {V}) Dijkstra's algorithm.

Do pruning on heap entry pushing: https://www.slideshare.net/yosupo/ss-46612984 P15

dfsPathSG :: Unindex i => SparseGraph i w -> i -> i -> Maybe [Vertex]

Returns a path from the source to the sink in reverse order. Note that it is NOT not the shortest path:

>>> reverse <$> dfsPathSG (buildUSG (0 :: Int, 3 :: Int) (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 3
Just [0,1,2,3]

treeDfsPathSG :: (HasCallStack, Unindex i) => SparseGraph i w -> i -> i -> [Vertex]

Returns a path from the source to the sink in reverse order. Note that it is NOT not the shortest path:

>>> reverse $ treeDfsPathSG (buildUSG (0 :: Int, 3 :: Int) (G.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 3
[0,1,2,3]

createDfsTreeSG :: Unindex i => SparseGraph i w -> i -> Vector Vertex

O(V+E) depth-first search. Returns a vector of parents. The source vertex or unrechable vertices are given `-1` as their parent.

>>> createBfsTreeSG (buildUSG (0 :: Int, 3 :: Int) (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0
[-1,0,1,1]

Retrieve a shortest path: >>> let ps = createBfsTreeSG (buildUSG (0 :: Int, 3 :: Int) (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 >>> restoreParentTreePath ps 3 [0,1,3]

createBfsTreeSG :: Unindex i => SparseGraph i w -> i -> Vector Vertex

O(V+E) breadth-first search. Returns a vector of parents. The source vertex or unrechable vertices are given `-1` as their parent.

>>> createBfsTreeSG (buildUSG (0 :: Int, 3 :: Int) (G.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0
[-1,0,1,1]

restoreParentTreePath :: Vector Vertex -> Vertex -> [Vertex]

Given a vector of vertex parents, restores path from the source to a sink.

topSortSG :: SparseGraph i w -> [Vertex]

Topological sort

Non-referenced vertices come first: >>> let !gr = buildUSG ((0, 4) :: (Int, Int)) $ U.fromList ([(0, 1), (0, 2), (2, 3)] :: [(Int, Int)]) >>> topSortSG gr [4,0,2,3,1]

topScc1SG :: forall i w m. PrimMonad m => SparseGraph i w -> MVector (PrimState m) Bool -> Vertex -> m [Vertex]

Partial running of scc over topologically sorted vertices, but for some connected components only.

revSG :: (Unindex i, Unbox w) => SparseGraph i w -> SparseGraph i w

Creates a reverse graph. TODO: return weightned graph

topSccSG :: (Unindex i, Unbox w) => SparseGraph i w -> [[Int]]

Collectes strongly connected components, topologically sorted. Upstream vertices come first, e.g., (v1 - v2) -> v3 -> v4.

treeDepthInfoSG :: SparseGraph Int w -> Int -> (ToParent, Vector Int)

LCA component. See also lca and lcaLen from Tree.

lcaCacheSG :: SparseGraph Int w -> Vertex -> LcaCache

LCA component. Returns LcaCache, i.e., `(parents, depths, parents')`.

bfs01_grid4_typical043 :: HasCallStack => IxUVector (Int, Int) Bool -> (Int, Int) -> IxUVector (Int, Int, Int) Int

01-BFS: zero cost with same direction.

bfsGrid317E_MBuffer :: HasCallStack => IxUVector (Int, Int) Bool -> (Int, Int) -> IxUVector (Int, Int) Int

TODO: Re-implement with genericBfs