Data.Tree.Fold

Description

Tree folding.

Documentation

foldTree :: forall m a. Array Vertex [Vertex] -> Vertex -> (m -> a -> a) -> (Vertex -> a) -> (a -> m) -> a

Folds a tree from one root vertex using postorder DFS.

scanTreeVG :: Vector v a => Array Vertex [Vertex] -> Vertex -> (m -> a -> a) -> (Vertex -> a) -> (a -> m) -> v a

Folds a tree from one root vertex using postorder DFS, recording all the accumulation values on every vertex.

scanTreeVU :: Unbox a => Array Vertex [Vertex] -> Vertex -> (m -> a -> a) -> (Vertex -> a) -> (a -> m) -> Vector a

Type-restricted scanTreeVG.

scanTreeV :: Array Vertex [Vertex] -> Vertex -> (m -> a -> a) -> (Vertex -> a) -> (a -> m) -> Vector a

Type-restricted scanTreeVG.

foldTreeAll :: (Unbox a, Unbox m, MonoidAction m a) => Array Vertex [Vertex] -> (Vertex -> a) -> (a -> m) -> Vector a

\(O(N)\). Folds a tree for every vertex as a root using the rerooting technique. REMARK: mempty is used for initial operator value.

Typical problems

Typical 039 - Tree Distance (★5)