Data.Tree.Lca

Documentation

newtype ToParent

Vector for retrieving the parent vertex.

Constructors

ToParent (Vector Vertex)

type LcaCache = (ToParent, Vector Int, BinaryLifting Vector ToParent)

treeDepthInfo :: Int -> (Int -> [Int]) -> Int -> (ToParent, Vector Int)

Returns `(parents, depths)` who maps vertices to the corresponding information. REMARK: Use 0-based index for the graph vertices. TODO: Consider using `Maybe Int` instead for easier Monoid integration

lcaCache :: Int -> (Vertex -> [Vertex]) -> Vertex -> LcaCache

Returns LcaCache, i.e., `(parents, depths, parents')`.

lca :: LcaCache -> Int -> Int -> (Int, Int)

Returns the lowest common ancestor `(v, d)` with the help of the binary lifting technique. REMARK: Use 0-based index for the graph vertices.

lcaLen :: LcaCache -> Int -> Int -> Int

Gets the length between given two vertices with the help of LCA.

newtype ToParentM m

ToParent with monoid concatanation.

Constructors

ToParentM (Int, Vector m)

type FoldLcaCache m = (LcaCache, Vector (Vector m))

LcaCache with monoid folding on path.

foldLcaCache :: forall m. (Monoid m, Unbox m) => Int -> (Vertex -> [Vertex]) -> Vertex -> (Vertex -> Vertex -> m) -> FoldLcaCache m

Returns FoldLcaCache that can be used for calculating the folding value of path between two vertices.

graph: Vertex -> [Vertex]

edgeValueOf: child -> parent -> m

foldLcaCache2 :: forall a m. (Monoid m, Unbox m) => Array Int [(Vertex, a)] -> (a -> m) -> FoldLcaCache m

foldLcaCache specific for `Array Vertex [(Vertex, a)]`.

foldViaLca :: forall m. (Monoid m, Unbox m) => FoldLcaCache m -> Int -> Int -> m

Calculates the folding value of the path between two vertices in a tree.