Data.Vector.Extra

Documentation

compressU :: HasCallStack => Vector Int -> (Vector Int, Vector Int)

\(O(N \log N)\) One dimensional index compression: xs -> (nubSortXs, xs')

bindex :: (HasCallStack, Vector v a, Ord a) => v a -> a -> Int

\(O(\log N)\) Binary search-based indexing.

chunksOfG :: Vector v a => Int -> v a -> Vector (v a)

\(O(N)\) = Test >>> chunksOfG 3 $ U.fromList ([1, 2, 3, 4, 5, 6, 7] :: [Int]) [[1,2,3],[4,5,6],[7]]

slideMinIndicesOn :: (Vector v a, Ord b) => (a -> b) -> Int -> v a -> Vector Int

\(O(N)\) https://qiita.com/kuuso1/items/318d42cd089a49eeb332

Returns indices of elements of the minimum values in corresponding spans. Drop the first len - 1 elements if you don't need spans shorter than the len.

slideMinIndices :: Int -> Vector Int -> Vector Int

\(O(N)\)

>>> slideMinIndices 3 (U.fromList [0 .. 5])
[0,1,2,3,4,5]
>>> slideMinIndices 3 (U.fromList [5, 4 .. 0])
[0,0,0,1,2,3]

slideMaxIndices :: Int -> Vector Int -> Vector Int

\(O(N)\)

>>> slideMaxIndices 3 (U.fromList [0 .. 5])
[0,0,0,1,2,3]
>>> slideMaxIndices 3 (U.fromList [5, 4 .. 0])
[0,1,2,3,4,5]

constructMN :: forall a m. (PrimMonad m, Unbox a) => Int -> (Vector a -> m a) -> m (Vector a)

\(O(N f)\)

constructrMN :: forall a m. (PrimMonad m, Unbox a) => Int -> (Vector a -> m a) -> m (Vector a)

\(O(N f)\)

prevPermutation :: (Ord e, Vector v e, Vector v (Down e)) => v e -> v e

invNum :: HasCallStack => Int -> Vector v Int => v Int -> Int

\(O(N \log N)\) Calculates the inversion number. Be sure to compress the input vector!

Typical problems

TODO

ABC 261 - F

compressInvNumG :: HasCallStack => Vector Int -> Int

\(O(N \log N)\) Calculates the inversion number, but after applying index compression. It can significantly improve the performance, like in ABC 261 F.

lexOrderMod :: (HasCallStack, Vector v Int) => v Int -> Int -> Int

\(O(N \log N)\) Returns 1-based lexicographic order of the given array.

WARNING: Use 0-based indices for the input.

Typical problems

TODO