Data.WaveletMatrix.SuccinctDictionary

Description

Succinct indexable dictionary functions.

Documentation

wordWM :: Int

Block sized for the internal cumultaive sum.

csumInPlaceBV :: PrimMonad m => MVector (PrimState m) Int -> Vector Bit -> m Int

\(O(N)\) Calculates the cumulative sum by word for the bit vector in-place.

csumBV :: Vector Bit -> Vector Int

\(O(N)\) Calculates the cumulative sum by word for the bit vector.

freq0BV :: Vector Bit -> Vector Int -> Int -> Int

\(O(1)\) Counts the number of 0 bits in interval [0, i). The input is a cumulative sum of bit vector by word. Rank 0.

freq1BV :: Vector Bit -> Vector Int -> Int -> Int

\(O(1)\) Counts the number of 1 bits in interval [0, i). The input is a cumulative sum of bit vector by word. Rank 1.

findKthIndex0BV :: Vector Bit -> Vector Int -> Int -> Maybe Int

\(O(\log N)\) Finds the index of kth 0. Select 0.

findKthIndex1BV :: Vector Bit -> Vector Int -> Int -> Maybe Int

\(O(\log N)\) Finds the index of kth 1. Select 1.