ToyLib.Debug.STree

Description

Debug utilities

Documentation

dbgSTree :: (Show (v a), Vector v a, PrimMonad m) => SegmentTree (Mutable v) (PrimState m) a -> m ()

Shows the leaves of a strict segment tree.

WARNING: It shows unused leaves, too.

dbgSTreeAll :: (Show (v a), Vector v a, PrimMonad m) => SegmentTree (Mutable v) (PrimState m) a -> m ()

Shows the nodes and the leaves of a strict segment tree,

WARNING: It shows unused nodes and leaves, too.

dbgLSTree :: (Show a, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> m ()

Shows the leaves of a lazily propagated segment tree.