ToyLib.IO

Description

IO

Primitives are ReadBS. Tuples of ReadBS are also ReadBS:

>>> :{
convertBS @(Int, Char, String, Float) $ BS.pack "42 c string 2.5"
:}
(42,'c',"string",2.5)

Vectors are ReadBS and they can also be embedded in the end of a tuple:

>>> :{
convertBS $ BS.pack "1 string 3.5 10 20 30 40" :: (Int, String, Float, VU.Vector Int)
:}
(1,"string",3.5,[10,20,30,40])

Documentation

int :: IO Int

ints :: IO [Int]

Reads one line as a list of integers.

class ReadBS a where

Read from a space-delimited ByteStrtig.

Minimal complete definition

Nothing

Methods

convertBS :: ByteString -> a

default convertBS :: Read a => ByteString -> a

readBS :: ByteString -> (a, ByteString)

For use with unfoldrExactN.

readMayBS :: ByteString -> Maybe (a, ByteString)

For use with unfoldr.

convertVG :: (ReadBS a, Vector v a) => ByteString -> v a

Converrts the given ByteString as a vector of a.

convertV :: ReadBS a => ByteString -> Vector a

convertVU :: (ReadBS a, Unbox a) => ByteString -> Vector a

convertNVG :: (ReadBS a, Vector v a) => Int -> ByteString -> v a

Converts the given ByteString as a vector of a with n elements.

convertNV :: ReadBS a => Int -> ByteString -> Vector a

convertNVU :: (ReadBS a, Unbox a) => Int -> ByteString -> Vector a

get :: ReadBS a => IO a

ints1 :: IO Int

ints2 :: IO (Int, Int)

ints3 :: IO (Int, Int, Int)

ints4 :: IO (Int, Int, Int, Int)

ints5 :: IO (Int, Int, Int, Int, Int)

ints6 :: IO (Int, Int, Int, Int, Int, Int)

intsVG :: Vector v Int => IO (v Int)

Reads one line as an integer.

intsV :: IO (Vector Int)

intsVU :: IO (Vector Int)

Reads one line as a vector of integers.

digitsVU :: IO (Vector Int)

intsGrid :: Int -> Int -> IO (IxVector (Int, Int) (Vector Int))

intsRestVG :: Vector v Int => IO (v Int)

intsRestVU :: IO (Vector Int)

getGraph :: Int -> Int -> IO (Array Int [Int])

Creates a graph from 1-based vertices

concat2 :: [(a, a)] -> [a]

concat two-item tuples

concatMap2 :: (a -> (b, b)) -> [a] -> [b]

getWGraph :: Int -> Int -> IO (Array Int [Entry Int Int])

Creates a weightend graph from 1-based vertices

getWGraph0 :: Int -> Int -> IO (Array Int [Entry Int Int])

Creates a weightend graph from 1-based vertices

endlBSB :: Builder

putBSB :: Builder -> IO ()

putLnBSB :: Builder -> IO ()

class ShowBSB a where

Show as a bytestring builder.

Minimal complete definition

Nothing

Methods

showBSB :: a -> Builder

default showBSB :: Show a => a -> Builder

showLnBSB :: ShowBSB a => a -> Builder

printBSB :: ShowBSB a => a -> IO ()

concatBSB :: Vector v a => (a -> Builder) -> v a -> Builder

See unwordsBSB as example.

unwordsBSB :: (ShowBSB a, Vector v a) => v a -> Builder

unlinesBSB :: (ShowBSB a, Vector v a) => v a -> Builder

yn :: Bool -> String

ynBSB :: Bool -> Builder

printYn :: Bool -> IO ()