ToyLib.Parser

Description

StateT -based parser.

TODO: Replace my IO module.

TODO: try strict tuple evaluation and compare performacne (force?)

Documentation

runIO :: StateT ByteString IO a -> IO a

Reads the whole stdin and runs the user program.

runFileIO :: StateT ByteString IO a -> String -> IO a

Reads a file runs the user program.

Primitives

int' :: MonadState ByteString m => m Int

Parses an Int.

int1' :: MonadState ByteString m => m Int

Parses an Int and substracts one.

char' :: MonadState ByteString m => m Char

word' :: MonadState ByteString m => m ByteString

double' :: MonadState ByteString m => m Double

Parsers an Int and converts it into a Double.

TODO: read and word'

Tuples

ints2' :: MonadState ByteString m => m (Int, Int)

ints11' :: MonadState ByteString m => m (Int, Int)

ints3' :: MonadState ByteString m => m (Int, Int, Int)

ints110' :: MonadState ByteString m => m (Int, Int, Int)

ints4' :: MonadState ByteString m => m (Int, Int, Int, Int)

ints5' :: MonadState ByteString m => m (Int, Int, Int, Int, Int)

ints6' :: MonadState ByteString m => m (Int, Int, Int, Int, Int, Int)

Readers

getLine' :: MonadState ByteString m => m ByteString

Reads one line from the state.

withLine' :: MonadState ByteString m => State ByteString a -> m a

Reads one line from the state and runs a pure parser for it.

More

auto' :: (ReadBS a, MonadState ByteString m) => m a

Parses one line via the ReadBS class.

intsU' :: MonadState ByteString m => m (Vector Int)

Reads an unboxed vector.

intsN' :: MonadState ByteString m => Int -> m (Vector Int)

Reads an unboxed vector with size n.

digitsU' :: MonadState ByteString m => m (Vector Int)

getHW' :: (Unbox a, ReadBS a, MonadState ByteString m) => Int -> Int -> m (Vector a)

Reads h lines of stdin and converts them as HxW **whitespace-delimited ByteString** and converts them into a flat vector of type a.

getMat' :: MonadState ByteString m => Int -> Int -> m (IxVector (Int, Int) (Vector Int))

Reads next h * w elements as matrix of type a.

getGrid' :: MonadState ByteString m => Int -> Int -> m (IxUVector (Int, Int) Char)

Reads next h * w elements as a char-based grid.