Algorithm.Bisect

Description

\(O(\log N)\) bisection method for sorted items in an inclusive range (from left to right only).

bisect returns an (ok, ng) index pair at the boundary. bisectL and bisectR returns one of the pair. bisectM is a monadic variant of bisect.

Example

With an OK predicate (<= 5), list [0..9] can be seen as:

[0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
 <-------------->  <-------->
        ok             ng

In the preceding example, bisect returns the (ok, ng) = (Just 5, Just 6) pair at the boundary:

>>> :{
let xs = [0 :: Int .. 9]
 in bisect (0 :: Int, 9 :: Int) (\i -> xs !! i <= 5)
:}
(Just 5,Just 6)

bisectL returns Just 5 and bisectR returns Just 6.

Bisection methods

Int

bisect :: Int -> Int -> (Int -> Bool) -> (Maybe Int, Maybe Int)

\(O(f \log N)\) Pure binary search.

bisectL :: Int -> Int -> (Int -> Bool) -> Maybe Int

\(O(f \log N)\) Also known as lower bound.

bisectR :: Int -> Int -> (Int -> Bool) -> Maybe Int

\(O(f \log N)\) Also known as upper bound.

bisectM :: forall m. Monad m => Int -> Int -> (Int -> m Bool) -> m (Maybe Int, Maybe Int)

\(O(f \log N)\) Monadic binary search.

bisectML :: forall m. Monad m => Int -> Int -> (Int -> m Bool) -> m (Maybe Int)

\(O(f \log N)\)

bisectMR :: forall m. Monad m => Int -> Int -> (Int -> m Bool) -> m (Maybe Int)

\(O(f \log N)\)

Double

bisectF64 :: Double -> Double -> Double -> (Double -> Bool) -> (Maybe Double, Maybe Double)

\(O(f \log N)\)

bisectF64L :: Double -> Double -> Double -> (Double -> Bool) -> Maybe Double

\(O(f \log N)\)

bisectF64R :: Double -> Double -> Double -> (Double -> Bool) -> Maybe Double

\(O(f \log N)\)

Binary search over vector

bsearch :: Vector v a => v a -> (a -> Bool) -> (Maybe Int, Maybe Int)

\(O(f \log N)\) bisect over a vector.

bsearchL :: Vector v a => v a -> (a -> Bool) -> Maybe Int

\(O(f \log N)\) bisectL over a vector.

bsearchR :: Vector v a => v a -> (a -> Bool) -> Maybe Int

\(O(f \log N)\) bisectR over a vector.

bsearchExact :: (Vector v a, Ord b) => v a -> (a -> b) -> b -> Maybe Int

\(O(\log N)\) bsearchExact over a vector, searching for a specific value. FIXME: It's slower than bsearchL.

bsearchM :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> Bool) -> m (Maybe Int, Maybe Int)

\(O(f \log N)\) bisectM over a vector.

bsearchML :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> Bool) -> m (Maybe Int)

\(O(f \log N)\) bisectML over a vector.

bsearchMR :: (PrimMonad m, MVector v a) => v (PrimState m) a -> (a -> Bool) -> m (Maybe Int)

\(O(f \log N)\) bisectMR over a vector.

bsearchMExact :: (PrimMonad m, MVector v a, Ord b) => v (PrimState m) a -> (a -> b) -> b -> m (Maybe Int)

\(O(\log N)\) bsearchMExact over a vector, searching for a specific value. FIXME: It's slower than bsearchL.

Misc

isqrtSlow :: Int -> Int

Retrieves square root of an Int.