Data.BinaryLifting

Description

Binary lifting is a technique for calculating nth power of a semigroup element in a (big) constant time, or applying them to their semigroup action target.

The i-th element of the underlying vector of BinaryLifting stores \(s^{2^i}\), with which we can construct any of \(s^i\) ((0 <= i < 2^63)) in a big ((63)) constant time.

Documentation

newtype BinaryLifting v m

Storage of \(s^{2^i}\).

Constructors

BinaryLifting (v m)

newBinLift :: (Vector v s, Semigroup s) => s -> BinaryLifting v s

Calculates BinaryLifting of the given semigroup

newBinLiftV :: Semigroup s => s -> BinaryLifting Vector s

Calculates BinaryLifting of the given semigroup

newBinLiftVU :: (Semigroup s, Unbox s) => s -> BinaryLifting Vector s

Calculates BinaryLifting of the given semigroup

stimesBL :: (Semigroup s, Vector v s) => BinaryLifting v s -> s -> Int -> s

Binarily lifted version of stimesMonoid. WARNING: Usually sactBL is much cheaper for semigroup actions with a boxed type.

mtimesBL :: (Monoid m, Vector v m) => BinaryLifting v m -> Int -> m

Binarily lifted version of stimesMonoid. WARNING: Usually sactBL is much cheaper for semigroup actions with a boxed type.

sactBL :: (SemigroupAction s a, Vector v s) => BinaryLifting v s -> a -> Int -> a

Binarily lifted semigroup action application.

mactBL :: (MonoidAction m a, Vector v m) => BinaryLifting v m -> a -> Int -> a

Alias of sactBL for monoid action.

newDoubling :: (Vector v a, Vector v Int) => a -> (a -> a) -> v a

Old binary lifting implementation without typeclasses. TODO: Remove

newDoublingV :: a -> (a -> a) -> Vector a

Old binary lifting implementation without typeclasses. TODO: Remove

applyDoubling :: Vector v op => v op -> a -> (a -> op -> a) -> Int -> a

Old binary lifting implementation without typeclasses. TODO: Remove