Data.BinaryLifting

Description

Type class-based binary lifting.

Binary lifting is a technique for calculating nth power of a semigroup element in a (big) constant time, or applying them to their semigroup action target.

The i-th element of the underlying vector of BinaryLifting stores \(s^{2^i}\), with which we can construct any of \(s^i\) ((0 <= i < 2^63)) in a big (63) constant time.

Documentation

class BinaryLifting a where

Binary lifting.

Associated Types

type VecBL a

Methods

cacheBL :: a -> VecBL a

cacheBLU :: (Semigroup a, Unbox a) => a -> Vector a

cacheBLV :: Semigroup a => a -> Vector a

stimesBL :: (Semigroup a, Vector v a) => v a -> Int -> a -> a

mtimesBL :: (Monoid a, Vector v a) => v a -> Int -> a

sactBL :: (SemigroupAction a b, Vector v a) => v a -> Int -> b -> b

type Permutation = TransiteSemigroup ()

idPerm :: Int -> Permutation

The identity element of Permutation.

newtype TransiteSemigroup a

Transition + semigroup concatanation. LCA is based on this.

Typical problems (Permutation)

Typical 058 - Original Calculator (★4) = Typical problems (Permutation)

Typical problems (TransiteSemigroup)

c.f. LCA.

Constructors

TransiteSemigroup (Vector (Int, a))

unTransiteSemigroup :: TransiteSemigroup a -> Vector (Int, a)

idTransiteSemigroup :: Unbox a => Int -> a -> TransiteSemigroup a

The identity element of TransiteSemigroup.

Orphan instances

SemigroupAction (Product Int) Int

Instance details

Methods

sact :: Product Int -> Int -> Int