Data.Graph.Generic

Description

Generic graph search.

Documentation

componentsOf :: (Vertex -> Vector Vertex) -> Int -> Vertex -> Vector Vertex

\(O(V+E)\) DFS that paints connnected components starting from a vertex.

genericBfs :: (Int -> Vector Int) -> Int -> Vertex -> Vector Int

\(O(V+E)\) breadth-first search. Unreachable vertices are given distance of -1.

genericBfs01 :: (Ix i, Unbox i) => (i, i) -> (i -> Vector (i, Int)) -> Int -> Vector i -> IxUVector i Int

\(O(V+E)\) 01-BFS. Unreachable vertices are given distance of -1.

Example

let !bnd3 = zero3 n n 4
let !res = genericBfs01 bnd3 grF (4 * n * n) source
      where
        grF (!y, !x, !iDir) = flip U.imapMaybe diag4 $ i (!dy, !dx) -> do
          let (!y', !x') = (y + dy, x + dx)
          guard $ inRange (zero2 n n) (y', x') && gr @! (y', x') /= #
          return ((y', x', i), bool 1 0 (iDir == i))
        source = U.generate 4 $ (sy, sx,)

Typical problems

ABC 176 E - Wizard in Maze

ABC 246 E - Bishop 2

genericDj :: forall w. (Unbox w, Num w, Ord w) => (Int -> Vector (Int, w)) -> Int -> Int -> w -> Vector Vertex -> Vector w

\(O((E+V) \log {V})\) Dijkstra's algorithm.

Does pruning on heap entry pushing: https://www.slideshare.net/yosupo/ss-46612984 P15

Note that longest Dijkstra can be implemented by just using a max heap.

genericSparseDj :: forall w. (Unbox w, Num w, Ord w) => (Int -> Vector (Int, w)) -> w -> Vector Vertex -> IntMap w

\(O((E+V) \log {V})\) Dijkstra's algorithm with sparse heap.

TODO: test

distsNN :: (Unbox w, Num w, Ord w) => Int -> w -> Vector (Int, Int, w) -> IxUVector (Int, Int) w

\(O(V^3)\) Floyd-Warshall algorith. It uses max as relax operator and the second argument is usually like maxBound div 2.

It's strict about path connection and invalid paths are ignored.