Data.Graph.Scc

Description

Strongly connected components and topological sort.

Documentation

topSort :: Array Int [Int] -> [Int]

Topological sort implemented with postorder DFS.

Implementation note

Topological sort is for DAG, but internally it's used for scc where cyclic graph input can come.

topScc1 :: forall m. PrimMonad m => Array Int [Int] -> MVector (PrimState m) Bool -> Int -> m [Int]

Partial running of scc over topologically sorted vertices, but for some connected components only.

revGraph :: Array Int [Int] -> Array Int [Int]

Retrieves a reverse graph

topScc :: Array Int [Int] -> [[Int]]

Collectes strongly connected components, topologically sorted. Upstream vertices come first, e.g., (v1 - v2) -> v3 -> v4.

topSccCycles :: Array Int [Int] -> [[Int]]

Collects cycles using scc.

downScc :: Array Int [Int] -> [[Int]]

Collectes strongly connected components, reverse topologically sorted. Down stream vertices come first, e.g., v4 <- v3 <- (v2 - v1)