Data.Graph.Sparse

Description

vector-based sparse graph implementation. Heavily inspired by cojna/iota.

Documentation

data SparseGraph w

CSR (compressed sparse row) representation of a graph, weightened or unweightened.

Constructors

SparseGraph

Fields

nVertsSG :: !Int

Number of vertices.

nEdgesSG :: !Int

Number of edges.

offsetsSG :: !(Vector Int)

Maps Vector to the starting edge index.

adjacentsSG :: !(Vector Vertex)

Adjacent vertices sorted with starting vertex.

edgeWeightsSG :: !(Vector w)

Edge weight information.

buildSG :: Int -> Vector (Vertex, Vertex) -> SparseGraph ()

Builds an non-weightned SparseGraph.

buildWSG :: Unbox w => Int -> Vector (Vertex, Vertex, w) -> SparseGraph w

Builds a weightned SparseGraph.

adj :: SparseGraph w -> Vertex -> Vector Vertex

Retrieves adjacent vertices.

eAdj :: SparseGraph w -> Vertex -> Vector (EdgeId, Vertex)

Returns (EdgeId, Vertex) paris. Hardly used.

adjW :: Unbox w => SparseGraph w -> Vertex -> Vector (Vertex, w)

Retrieves adjacent vertices with weights.

dfsSG :: SparseGraph w -> Vertex -> Vector Int

\(O(V+E)\) Depth-first search. Returns a vector of distances to each vertex. Unreachable vertices are given distance of `-1`.

dfsEveryPathSG :: forall w. (Num w, Ord w, Unbox w) => SparseGraph w -> Vertex -> w

\(O(V+E)\) Depth-first search. Just a template.

Typical problems

ABC 317 C - Remembering the Days

componentsSG :: SparseGraph w -> Vector Vertex -> Vector Vertex

\(O(V+E)\) Collects reachable vertices from source points.

allComponentsSG :: SparseGraph w -> (Vector Int, [[Vertex]])

\(O(V+E)\) DFS that paints connnected components. Returns (vertexToComponentId, components). Works on a non-directed graph only.

bfsSG :: SparseGraph w -> Vertex -> Vector Int

\(O(V+E)\) Breadth-first search. Unreachable vertices are given distance of -1.

bfs01SG :: SparseGraph Int -> Vector Vertex -> Vector Int

\(O(V+E)\) 01-BFS. Unreachable vertices are given distance of -1.

djSG :: forall w. (Num w, Ord w, Unbox w) => SparseGraph w -> w -> Vector Vertex -> Vector w

\(O((E+V) \log {V})\) Dijkstra's algorithm.

>>> let gr = buildWSG 4 (U.fromList [(0, 1, 1 :: Int), (1, 2, 1), (1, 3, 100), (2, 3, 1)])
>>> djSG gr (-1 :: Int) (U.singleton 0)
[0,1,2,3]

dfsPathSG :: SparseGraph w -> Vertex -> Vertex -> Maybe [Vertex]

\(O(V+E)\) Returns a path from the source to the sink in reverse order. Note that it is NOT not the shortest path:

>>> reverse <$> dfsPathSG (buildSG 4 (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 3
Just [0,1,2,3]

treeDfsPathSG :: HasCallStack => SparseGraph w -> Vertex -> Vertex -> [Vertex]

\(O(V+E)\)Returns a path from the source to the sink in reverse order. Note that it is NOT not the shortest path:

>>> reverse $ treeDfsPathSG (buildSG 4 (G.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 3
[0,1,2,3]

createDfsTreeSG :: SparseGraph w -> Vertex -> Vector Vertex

\(O(V+E)\) depth-first search. Returns a vector of parents. The source vertex or unrechable vertices are given `-1` as their parent.

>>> createBfsTreeSG (buildSG 4 (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0
[-1,0,1,1]

Retrieve a shortest path: >>> let ps = createBfsTreeSG (buildSG 4 (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 >>> restorePath ps 3 [0,1,3]

createBfsTreeSG :: SparseGraph w -> Vertex -> Vector Vertex

\(O(V+E)\) breadth-first search. Returns a vector of parents. The source vertex or unrechable vertices are given `-1` as their parent.

>>> createBfsTreeSG (buildSG 4 (G.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0
[-1,0,1,1]

restorePath :: Vector Vertex -> Vertex -> Vector Vertex

\(O(V)\) Given a vector of vertex parents, restores path from the source to a sink.

TODO: restore without reverse?

data DigraphInfo

Tries to paint the whole graph (possible not connected) as a digraph.

let DigraphInfo isFailure vertColors vertComps compInfos = paintDigraphSG gr

Typical problems

ABC 282 D - Make Bipartite 2

PAST 10 O - 3-順列

Constructors

DigraphInfo

Fields

isAllDigraphDI :: Bool

False if any of the connected components is not a digraph.

vertColorDI :: Vector Int

Vertex -> color (0 or 1)

vertComponentDI :: Vector Int

Vertex -> component index

componentInfoDI :: Vector (Int, Int, Bool)

component -> (n1, n2, isDigraph)

digraphSG :: SparseGraph w -> DigraphInfo

\(O(V+E)\) Tries to paint the whole graph (possible not connected) as a digraph. Works on non-directed graphs only.

topSortSG :: SparseGraph w -> [Vertex]

\(O(V+E)\) Topological sort

Upstream (not referenced) vertices come first: >>> let !gr = buildSG 5 $ U.fromList ([(0, 1), (0, 2), (2, 3)] :: [(Int, Int)]) >>> topSortSG gr [4,0,2,3,1]

revSG :: Unbox w => SparseGraph w -> SparseGraph w

\(O(V+E)\) Creates a reverse graph.

topScc1SG :: forall w m. PrimMonad m => SparseGraph w -> MVector (PrimState m) Bool -> Vertex -> m [Vertex]

\(O(V+E)\) Partial running of topSccSG over topologically sorted vertices, but for some connected components only.

topSccSG :: Unbox w => SparseGraph w -> [[Int]]

\(O(V+E)\) Collectes strongly connected components, reverse topologically sorted.

Downstream vertices come first, e.g., [v1 == v2 == v3] <- [v4 == v5] <- [v6].

collectMST :: (Ord w, Unbox w) => Int -> Vector (Vertex, Vertex, w) -> Vector (Vertex, Vertex, w)

\(O(E)\) Kruscal's algorithm. Returns edges for building a minimum spanning tree.

NOTE: User ia assumed to not duplicate the edges.

If you need to detect if the tree spans all the vertices, be sure to check the length of the edges equals to the number of vertices minus one. Do not try to count the vertices directly, e.g., o-o o-o contains four connected vertices, but only has two edges. It's apparently not a spanning tree.

Typical problems

Typica 49 - Flip Digits 2

buildMST :: (Ord w, Unbox w) => Int -> Vector (Vertex, Vertex, w) -> SparseGraph w

\(O(E)\) Kruscal's algorithm. Returns a minimum spanning tree.

NOTE: User ia assumed to not duplicate the edges.