Data.Graph.Sparse

Description

vector-based sparse graph implementation. Heavily inspired by cojna/iota.

Documentation

data SparseGraph i w

CSR (compressed sparse row) representation of a graph, weightened or unweightened.

Constructors

SparseGraph

Fields

boundsSG :: !(i, i)

Vertex index boundary.

nVertsSG :: !Int

Number of vertices.

nEdgesSG :: !Int

Number of edges.

offsetsSG :: !(Vector Int)

Maps Vector to the starting edge index.

adjacentsSG :: !(Vector Vertex)

Adjacent vertices sorted with starting vertex.

edgeWeightsSG :: !(Vector w)

Edge weight information.

buildSG :: Unindex i => (i, i) -> Vector (i, i) -> SparseGraph i ()

Builds an non-weightned SparseGraph.

buildWSG :: (Unindex i, Unbox w) => (i, i) -> Vector (i, i, w) -> SparseGraph i w

Builds a weightned SparseGraph.

buildRawSG :: (Unindex i, Unbox w) => (i, i) -> Vector (Vertex, Vertex, w) -> SparseGraph i w

adj :: SparseGraph i w -> Vertex -> Vector Vertex

Retrieves adjacent vertices.

eAdj :: SparseGraph i w -> Vertex -> Vector (EdgeId, Vertex)

Returns (EdgeId, Vertex) paris. Hardly used.

adjIx :: Unindex i => SparseGraph i w -> i -> Vector i

Retrieves adjacent vertex indices.

adjW :: Unbox w => SparseGraph i w -> Vertex -> Vector (Vertex, w)

Retrieves adjacent vertices with weights.

adjIxW :: (Unindex i, Unbox w) => SparseGraph i w -> i -> Vector (i, w)

Retrieves adjacent vertex indices with weights.

dfsSG :: Unindex i => SparseGraph i w -> i -> IxVector i (Vector Int)

\(O(V+E)\) Depth-first search. Returns a vector of distances to each vertex. Unreachable vertices are given distance of `-1`.

dfsEveryPathSG :: SparseGraph Int Int -> Int -> Int

\(O(V+E)\) Depth-first search. Just a template.

Typical problems

ABC 317 C - Remembering the Days

nonGenericDfs :: PrimMonad m => Int -> (Vertex -> m (Vector Vertex)) -> (Vertex -> m ()) -> Vertex -> m ()

A side effect-only not generic-enough DFS.

componentsVecSG :: Ix i => SparseGraph i w -> i -> IxVector i (Vector Bool)

\(O(V+E)\) Marks connected vertices. Also consider using union-find tree.

bfsSG :: Ix i => SparseGraph i w -> i -> IxVector i (Vector Int)

\(O(V+E)\) breadth-first search. Unreachable vertices are given distance of -1.

genericBfs :: (Int -> Vector Int) -> Int -> Vertex -> Vector Int

\(O(V+E)\) breadth-first search. Unreachable vertices are given distance of -1.

genericBfs01 :: (Ix i, Unbox i) => (i, i) -> (i -> Vector (i, Int)) -> Int -> Vector i -> IxUVector i Int

\(O(V+E)\) 01-BFS. Unreachable vertices are given distance of -1.

djSG :: forall i w. (Ix i, Unbox i, Num w, Ord w, Unbox w) => SparseGraph i w -> w -> Vector i -> IxUVector i w

\(O((E+V) \log {V})\) Dijkstra's algorithm.

>>> let gr = buildWSG (0 :: Int, 3 :: Int) (U.fromList [(0, 1, 1 :: Int), (1, 2, 1), (1, 3, 100), (2, 3, 1)])
>>> vecIV $ djSG gr (-1 :: Int) (U.singleton 0)
[0,1,2,3]

genericDj :: forall w. (Unbox w, Num w, Ord w) => (Int -> Vector (Int, w)) -> Int -> Int -> w -> Vector Vertex -> Vector w

\(O((E+V) \log {V})\) Dijkstra's algorithm.

Does pruning on heap entry pushing: https://www.slideshare.net/yosupo/ss-46612984 P15

Note that longest Dijkstra can be implemented by just using a max heap.

genericSparseDj :: forall w. (Unbox w, Num w, Ord w) => (Int -> Vector (Int, w)) -> w -> Vector Vertex -> IntMap w

dfsPathSG :: Unindex i => SparseGraph i w -> i -> i -> Maybe [Vertex]

Returns a path from the source to the sink in reverse order. Note that it is NOT not the shortest path:

>>> reverse <$> dfsPathSG (buildSG (0 :: Int, 3 :: Int) (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 3
Just [0,1,2,3]

treeDfsPathSG :: (HasCallStack, Unindex i) => SparseGraph i w -> i -> i -> [Vertex]

Returns a path from the source to the sink in reverse order. Note that it is NOT not the shortest path:

>>> reverse $ treeDfsPathSG (buildSG (0 :: Int, 3 :: Int) (G.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 3
[0,1,2,3]

createDfsTreeSG :: Unindex i => SparseGraph i w -> i -> Vector Vertex

\(O(V+E)\) depth-first search. Returns a vector of parents. The source vertex or unrechable vertices are given `-1` as their parent.

>>> createBfsTreeSG (buildSG (0 :: Int, 3 :: Int) (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0
[-1,0,1,1]

Retrieve a shortest path: >>> let ps = createBfsTreeSG (buildSG (0 :: Int, 3 :: Int) (U.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0 >>> restorePath ps 3 [0,1,3]

createBfsTreeSG :: Unindex i => SparseGraph i w -> i -> Vector Vertex

\(O(V+E)\) breadth-first search. Returns a vector of parents. The source vertex or unrechable vertices are given `-1` as their parent.

>>> createBfsTreeSG (buildSG (0 :: Int, 3 :: Int) (G.fromList [(0, 1), (1, 2), (1, 3), (2, 3)])) 0
[-1,0,1,1]

restorePath :: Vector Vertex -> Vertex -> Vector Vertex

Given a vector of vertex parents, restores path from the source to a sink.

TODO: restore without reverse?

topSortSG :: SparseGraph i w -> [Vertex]

Topological sort

Non-referenced vertices come first: >>> let !gr = buildSG ((0, 4) :: (Int, Int)) $ U.fromList ([(0, 1), (0, 2), (2, 3)] :: [(Int, Int)]) >>> topSortSG gr [4,0,2,3,1]

revTopScc1SG :: forall i w m. PrimMonad m => SparseGraph i w -> MVector (PrimState m) Bool -> Vertex -> m [Vertex]

Partial running of topSccSG over topologically sorted vertices, but for some connected components only.

revSG :: (Unindex i, Unbox w) => SparseGraph i w -> SparseGraph i w

Creates a reverse graph. TODO: return weightned graph

revTopSccSG :: (Unindex i, Unbox w) => SparseGraph i w -> [[Int]]

Collectes strongly connected components, reverse topologically sorted. Upstream vertices come first, e.g., v1 -> v2 -> v3.

topSccSG :: (Unindex i, Unbox w) => SparseGraph i w -> [[Int]]

Collectes strongly connected components, topologically sorted. Downstream vertices come first, e.g., v3 -> v2 -> v1.

treeDepthInfoSG :: (Monoid w, Unbox w) => SparseGraph Int w -> Int -> (Vector Int, TransiteSemigroup w)

Returns (depths, parents).

lcaCacheSG :: (Monoid w, Unbox w) => SparseGraph Int w -> Vertex -> LcaCache w

Returns LcaCache, i.e., `(parents, depths, parents')`.

foldTreeImpl :: forall m op a w. Monad m => SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> (Vertex -> a -> m ()) -> m a

foldTree :: SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> a

Folds a tree from one root vertex using postorder DFS.

scanTree :: Vector v a => SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> v a

Folds a tree from one root vertex using postorder DFS, recording all the accumulation values on every vertex.

scanTreeU :: Unbox a => SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> Vector a

Type-restricted scanTree.

scanTreeV :: SparseGraph Int w -> Vertex -> (op -> a -> a) -> (Vertex -> a) -> (a -> op) -> Vector a

Type-restricted scanTree.

foldTreeAllSG :: forall a op w. (Unbox a, Unbox op, MonoidAction op a) => SparseGraph Int w -> (Vertex -> a) -> (a -> op) -> Vector a

\(O(N)\). Folds a tree for every vertex as a root using the rerooting technique. REMARK: mempty is used for initial operator value.

Typical problems

Typical 039 - Tree Distance (★5)

EDPC V - Subtree

TDPC N - tree

collectMST :: (Ord w, Unbox w) => Int -> Vector (Int, Int, w) -> Vector (Int, Int, w)

\(O(E)\) Kruscal's algorithm. Returns used edges.

Typical problems

TODO: Test it

buildMST :: (Ord w, Unbox w) => Int -> Vector (Int, Int, w) -> SparseGraph Int w

distsNN :: (Unbox w, Num w, Ord w) => Int -> w -> Vector (Int, Int, w) -> IxUVector (Int, Int) w

\(O(V^3)\) Floyd-Warshall algorith. It uses max as relax operator and the second argument is usually like maxBound div 2.

It's strict about path connection and invalid paths are ignored.