Data.Graph.Tree.Hld

Description

Heavy-light decomposition. Heavily inspired by cojna/iota and maspypy/library.

HLD lets you find LCA in \(O(\log V)\) time and lets you fold monoids on a path in \(O(\log^2 V)\) time with the help of segment trees.

Documentation

type VertexHLD = Vertex

Vertex reindexed by indexHLD.

data HLD

HLD splits a tree into lines. Lines are often called "paths", but here we say "lines" for avoiding word conflicts.

Example

Original tree

Consider a tree with arbitrary vertex order:

 0==8==7==3==1==2==12==13==15==15     XX: vertex
    |        |                         ==: edge on the same line
10==5        11==8==6                  --: edge between different lines
    |
    4

indexHLD visualized

The tree vertices are reindexed with indexHLD, ensuring vertices in each line have consecutive numbers:

 0==1==2==3==4==5==6==7==8==9
    |        |
14==13       10==11==12
    |
    15

Note that vertices on higher lines are assigned smaller numbers.

headHLD visualized

headHLD can be used for finding LCA of two vertices. To find the LCA, move up to the head, go up to the parental line's vertex and repeat until they are on the same line.

 0==0==0==0==0==0==0==0==0==0
    |     |
13==13     10==10==10
    |
    15

headHLD also works for identifying lines. When the two vertices are on the same line, they have the same head.

parentHLD visualized

parentHLD lets you go up to the parental line's vertex from a head:

(-1)==0==1==2==3==4==5==6==-7==8
      |        |
  13==1        4==10==11
      |
     13

Segment tree integration

With HLD, we can find a path between two vertices u, v: u -> lca(u, v) -> v. The path is composed of line slices, which can be fastly folded with a segment tree.

Vertex monoid folding on a path

If vertices have weights, use foldVertsCommuteHLD or foldVertsHLD.

Edge monoid folding on a path

If edges have weights, you can either treat the edges as new vertices or put weight to the depper index.

Idea 1. Treat edges as new vertices.

o--o--o  -> o-x-o-x-o

Idea 2. Assign edge weight to the deeper vertex. This is the idea of foldEdgesCommuteHLD and foldEdgesHLD:

  o
  | <--- edge 1
  o <- write wegiht 1 here
  | <--- edge 2
  o <- write wegiht 2 here

Constructors

HLD

Fields

parentHLD :: !(Vector Vertex)

Vertex -> Parent Vertex.

indexHLD :: !(Vector VertexHLD)

Vertex -> Reindexed vertex (VertexHLD).

headHLD :: !(Vector Vertex)

Vertex -> The line's head Vertex.

Construction

hldOf :: forall w. SparseGraph Int w -> HLD

\(O(V)\) Heavy-light decomposition or Centroid Path Decomposition construction.

LCA and paths

lcaHLD :: HLD -> Vertex -> Vertex -> Vertex

\(O(\log V)\) HLD calculation.

Typical Problems

ABC 133 - F

_pathHLD :: Bool -> HLD -> Vertex -> Vertex -> [(VertexHLD, VertexHLD)]

\(O(\log V)\) Shared implementation of edgePathHLD and vertPathHLD.

edgePathHLD :: HLD -> Vertex -> Vertex -> [(VertexHLD, VertexHLD)]

\(O(\log V)\) Returns inclusive edge vertex pairs.

vertPathHLD :: HLD -> Vertex -> Vertex -> [(VertexHLD, VertexHLD)]

\(O(\log V)\) Returns inclusive vertex pairs per HLD line.

Folding methods

_foldHLD :: (Monoid mono, Monad m) => Bool -> HLD -> (VertexHLD -> VertexHLD -> m mono) -> (VertexHLD -> VertexHLD -> m mono) -> Vertex -> Vertex -> m mono

\(O(\log V)\) The common implementation of foldEdges*HLD variants.

foldEdgesCommuteHLD :: (Monoid mono, Monad m) => HLD -> (VertexHLD -> VertexHLD -> m mono) -> Vertex -> Vertex -> m mono

\(O(\log V)\) Folds commutative monoids on a tree edges using HLD. Prefer to use the wrapper TreeMonoid.

Typical Problems

ABC 294 - G

foldEdgesHLD :: (Monoid mono, Monad m) => HLD -> (VertexHLD -> VertexHLD -> m mono) -> (VertexHLD -> VertexHLD -> m mono) -> Vertex -> Vertex -> m mono

Folds commutative monoids on a tree vertices using HLD. Prefer to use the wrapper TreeMonoid.

TODO: verify

foldVertsCommuteHLD :: (Monoid mono, Monad m) => HLD -> (VertexHLD -> VertexHLD -> m mono) -> Vertex -> Vertex -> m mono

Folds commutative monoids on a tree vertices using HLD. Prefer to use the wrapper TreeMonoid.

Typical Problems

Vertex Add Path Sum - Library Checker

foldVertsHLD :: (Monoid mono, Monad m) => HLD -> (VertexHLD -> VertexHLD -> m mono) -> (VertexHLD -> VertexHLD -> m mono) -> Vertex -> Vertex -> m mono

Folds non-commutative monoids on a tree vertices using HLD. Prefer to use the wrapper TreeMonoid.

Typical Problems

Vertex Set Path Composite - Library Checker

TreeMonoid

data TreeMonoid a s

HLD wrapper for folding paths on a tree using HLD and segment tree(s).

Constructors

TreeMonoid

Fields

hldTM :: !HLD

Borrowed HLD.

isCommuteTM :: !Bool

Indicates if it's targetting commutative monoids.

isEdgeTM :: !Bool

Indicates if it's targetting edge weights. (If not, it's targetting vertex weights on no).

streeFTM :: !(SegmentTree MVector s a)

Segment tree for folding upwards.

streeBTM :: !(SegmentTree MVector s (Dual a))

Segment tree for folding downwards. Only created when the monoid is not commutative.

Construction

_buildRawTM :: (PrimMonad m, Monoid a, Unbox a) => HLD -> Bool -> Bool -> Vector a -> m (TreeMonoid a (PrimState m))

\(O(V)\) Shared implementation of buildVertTM and buildEdgeTM.

buildVertTM :: (PrimMonad m, Monoid a, Unbox a) => HLD -> Bool -> Vector a -> m (TreeMonoid a (PrimState m))

\(O(V)\) Builds a TreeMonoid on vertices.

edgeVertsHLD :: Unbox a => HLD -> Vector (Vertex, Vertex, a) -> Vector (Vertex, a)

\(O(V)\) Map weightened edges into (Vertex, a) pairs. The output is the input to buildEdgeTM.

buildEdgeTM :: (PrimMonad m, Monoid a, Unbox a) => HLD -> Bool -> Vector (Vertex, a) -> m (TreeMonoid a (PrimState m))

\(O(V)\) Builds a TreeMonoid on edges. **The input must be the output of edgeVertsHLD**.

Segment tree methods

foldTM :: (PrimMonad m, Monoid a, Unbox a) => TreeMonoid a (PrimState m) -> Vertex -> Vertex -> m a

\(O(log^2 V)\) Folds TreeMonoid on a path between two vertices

readTM :: (PrimMonad m, Unbox a) => TreeMonoid a (PrimState m) -> Vertex -> m a

\(O(log V)\) Reads a TreeMonoid value on a Vertex.

writeTM :: (PrimMonad m, Monoid a, Unbox a) => TreeMonoid a (PrimState m) -> Vertex -> a -> m ()

\(O(log V)\) Write a TreeMonoid value on a Vertex.

exchangeTM :: (PrimMonad m, Monoid a, Unbox a) => TreeMonoid a (PrimState m) -> Vertex -> a -> m a

\(O(log V)\) Exchanges a TreeMonoid value on a Vertex.

modifyTM :: (PrimMonad m, Monoid a, Unbox a) => TreeMonoid a (PrimState m) -> (a -> a) -> Int -> m ()

\(O(log V)\) Modifies a TreeMonoid value on a Vertex.