Data.RollingHash

Description

The rolling hash algorithm lets you create ( \(O(N)\) or \(O(N \log N)\) ) fastly ( \(O(1)\) ) comparable, concatanatable string slices.

Overview

Rolling hash of a string is calculated as a B-adic number. As an example, slice [2, 4] of a string "abcde" is given as this:

           s :=     a       b       c       d       e
           s4 = b^4 a + b^3 b + b^2 c + b^1 d + b^0 e
           s2 = b^1 a + b^0 b
s4 - s2 * b^3 =                 b^2 c + b^1 d + b^0 e

Cumulative sum-based implementation

RollingHash is a cumulative sum-based implementation. It needs to calculate \(b^{n_1} / b^{n_2}\) for the inverse operation. Thus it holds a cache of \(\{b^{n}\}_{n}\) and not so flexible. Unless extreamly fast calculaton is required, RH is better.

Monoid-based implmentation

RH is the monoid-based rolling hash implementation. Convert a string into a segment tree of RH and you win.

Documentation

data RH b p

Rolling hash monoid. Essensially a (Dual) Affine2d ModInt.

Typical prolbems

ABC 331 F - Palindrome Query

Constructors

RH

Fields

nextDigitRH :: !Int

\(b^{length}\).

hashRH :: !Int

The hash value.

rh1 :: forall b p. KnownNat b => Int -> RH b p

\(O(1)\) Creates a one-length RH from an integer.

Warning

The input must be less than p.

type RHRepr = A2 Int

data RollingHash b p

Rolling hash of a string.

Internals

Slice (2, 4) of "abcdef" is given as this:

           s :=     a       b       c       d       e
           s4 = b^4 a + b^3 b + b^2 c + b^1 d + b^0 e
           s2 = b^1 a + b^0 b
s4 - s2 * b^3 =                 b^2 c + b^1 d + b^0 e

Constructors

RollingHash

Fields

sourceLength :: !Int

dimensions :: !(Vector Int)

\(\{B^i mod p\}_{i \elem [0, n)}\)

hashSum :: !(Vector Int)

type HashInt = 100 :: Nat

Type level integer that represents the B-adic number used for the rolling hash algorithm.

newRH :: forall p. KnownNat p => String -> RollingHash HashInt p

\(O(N)\) Creates a rolling hash of given string.

lengthRH :: RollingHash b p -> Int

\(O(1)\) Retrieves the original length of the RollingHash string.

data HashSlice p

HashSlice value length. See also the example of RollingHash.

Constructors

HashSlice

Fields

hashValue :: !Int

hashLength :: !Int

sliceRH :: forall b p. KnownNat p => RollingHash b p -> Int -> Int -> HashSlice p

\(O(1)\) Slices a rolling hash string.

consHS :: forall b p. KnownNat p => RollingHash b p -> HashSlice p -> HashSlice p -> HashSlice p

\(O(1)\) Cons two rolling hash slices.

emptyHS :: HashSlice p

\(O(1)\) Creates an empty rolling hash slice.

concatHS :: forall b p t. (KnownNat p, Foldable t) => RollingHash b p -> t (HashSlice p) -> HashSlice p

\(O(K)\) Concatanates two rolling hash slices.