Data.SegmentTree.Lazy

Description

Lazy segment tree, where we can perform operation over range.

Typical problems

Typical 029 - Long Bricks (★5)

EDPC W - Intervals

TODO: Add set get, as in [ac-library](https:github.comatcoderac-libraryblobmasteratcoder/lazysegtree.hpp)

Documentation

data LazySegmentTree v a op s

Lazy segment tree.

Indices

Use 1-based indices for super handy index hacks:

           1             |
     2           3       | height = 4
  4     5     6     7    |
08 09 10 11 12 13 14 15  v
^
+-- nVerts / 2

0  1  2  3  4  5  6  7   -- iLeaf is given by user and uses zero-based indices.

- parentVertex = vertex / 2 = shiftR vertex 1
- leftChild = vertex * 2 = shiftR vertex 1
- rightChild = vertex * 2 + 1 = shiftR vertex 1 + 1 = (shiftR vertex 1) .|. 1

Invariant

New operators always come from right: oldOp <> newOp

Constructors

LazySegmentTree !(v s a) !(MVector s op) !Int

newLazySTree :: forall v a op m. (MVector v a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> m (LazySegmentTree v a op (PrimState m))

Creates LazySegmentTree with mempty as the initial accumulated values.

newLazySTreeV :: forall a op m. (Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> m (LazySegmentTree MVector a op (PrimState m))

newLazySTreeVU :: forall a op m. (Unbox a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> m (LazySegmentTree MVector a op (PrimState m))

generateLazySTree :: forall v a op m. (MVector v a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> (Int -> a) -> m (LazySegmentTree v a op (PrimState m))

Creates LazySegmentTree with initial leaf values. FIXME: Use mempty for $i >= n$.

generateLazySTreeV :: forall a op m. (Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> (Int -> a) -> m (LazySegmentTree MVector a op (PrimState m))

generateLazySTreeVU :: forall a op m. (Unbox a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> (Int -> a) -> m (LazySegmentTree MVector a op (PrimState m))

updateLazySTree :: forall v a op m. (MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> op -> m ()

Appends the lazy operator monoid monoids over some span of the lazy segment tree. These values are just stored and performed over the nodes when queried.

queryLazySTree :: forall v a m op. (MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> m a

_propOpMonoidsToLeaf :: (MVector v a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> m ()

Propagates the lazy operator monoids from top to bottom where the laef vertex is contained.

iLeaf: Given with zero-based index.

_evalToRoot :: (MVector v a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> m ()

Evaluates parent values on updateSegmentTree. TODO: move to where clause of the update function?