Data.SegmentTree.Lazy

Description

Lazily propageted segment tree, where we can perform operation over range.

Algebra

(Op * Acc) <> (Op * Acc) = (Op <> Op) (Acc <> Acc)

Typical problems

Typical 029 - Long Bricks (★5)

EDPC W - Intervals

TODO: Add set get, as in [ac-library](https:github.comatcoderac-libraryblobmasteratcoder/lazysegtree.hpp)

TODO: consider NodeAct naming rather than AccOp.

Documentation

data LazySegmentTree v a op s

Lazy segment tree.

(Internal) 1-based indices

Use 1-based indices for super handy vertex indices:

           1             |
     2           3       | height = 4 = log_2 16
  4     5     6     7    |
08 09 10 11 12 13 14 15  v
^
+-- nVerts / 2

0  1  2  3  4  5  6  7   -- iLeaf is given by user and uses zero-based indices.

- parent = v .>>. 1
- childL = v .<<. 1
- childR = v .<<. 1 .|. 1

Invariant

New operators always come from the left: (newOp <> oldOp) acc or newOp sact (oldOp sact ac).

Constructors

LazySegmentTree !(v s a) !(MVector s op) !Int

newLSTreeImpl :: forall a op v m. (Monoid a, MonoidAction op a, Unbox op, MVector v a, PrimMonad m) => Int -> m (LazySegmentTree v a op (PrimState m))

Creates LazySegmentTree with mempty as the initial accumulated values.

newLSTree :: forall a op m. (Unbox a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> m (LazySegmentTree MVector a op (PrimState m))

generateLSTreeImpl :: forall v a op m. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> (Int -> a) -> m (LazySegmentTree v a op (PrimState m))

Creates LazySegmentTree with initial leaf values.

generateLSTree :: forall a op m. (HasCallStack, Unbox a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => Int -> (Int -> a) -> m (LazySegmentTree MVector a op (PrimState m))

updateLSTree :: forall v a op m. (MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> op -> m ()

Appends the lazy operator monoid monoids over a span. They are just stored and propagated when queried.

foldLSTree :: forall v a m op. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> m a

readLSTree :: forall a op v m. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> m a

Read one leaf.

foldMayLSTree :: forall a op v m. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> m (Maybe a)

foldWholeLSTree :: forall a op v m. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> m a

_propOpMonoidsToLeaf :: (HasCallStack, MVector v a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> m ()

Propagates the lazy operator monoids from top to bottom where the leaf vertex is contained.

iLeaf: Given with zero-based index.

_evalToRoot :: (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> m ()

Evaluates parent values on updateSegmentTree. TODO: move to where clause of the update function?

bisectLSTree :: forall v a m op. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> (a -> Bool) -> m (Maybe Int, Maybe Int)

The l, r indices are the zero-based leaf indices.

bisectLSTreeL :: forall v a m op. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> (a -> Bool) -> m (Maybe Int)

bisectLSTreeR :: forall v a m op. (HasCallStack, MVector v a, Monoid a, MonoidAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree v a op (PrimState m) -> Int -> Int -> (a -> Bool) -> m (Maybe Int)