Data.SegmentTree.Lazy

Description

Lazily propageted segment tree, where we can perform operation over range.

Algebra

(Op * Acc) <> (Op * Acc) = (Op <> Op) * (Acc <> Acc)

Typical example of Op is a matrix and Acc is a column vector.

Typical problems

Typical 029 - Long Bricks (★5)

EDPC W - Intervals

Documentation

data LazySegmentTree a op s

Lazy segment tree.

(Internal) 1-based indices

Use 1-based indices for super handy vertex indices:

           1             |
     2           3       | height = 4 = log_2 16
  4     5     6     7    |
08 09 10 11 12 13 14 15  v
^
+-- nVerts / 2

0  1  2  3  4  5  6  7   -- iLeaf is given by user and uses zero-based indices.

- parent = v .>>. 1
- childL = v .<<. 1
- childR = v .<<. 1 .|. 1

Invariant

New operators always come from the left: (newOp <> oldOp) acc or newOp sact (oldOp sact ac).

Constructors

LazySegmentTree !(MVector s a) !(MVector s op) !Int

newLSTreeImpl :: forall a op m. (Monoid a, Unbox a, Monoid op, Unbox op, PrimMonad m) => Int -> m (LazySegmentTree a op (PrimState m))

\(O(N)\) Creates LazySegmentTree with mempty as the initial accumulated values.

newLSTree :: forall a op m. (Unbox a, Monoid a, Monoid op, Unbox op, PrimMonad m) => Int -> m (LazySegmentTree a op (PrimState m))

\(O(N)\)

generateLSTreeImpl :: forall a op m. (HasCallStack, Monoid a, Unbox a, Monoid op, Unbox op, PrimMonad m) => Int -> (Int -> a) -> m (LazySegmentTree a op (PrimState m))

\(O(N)\) Creates LazySegmentTree with initial leaf values.

generateLSTree :: forall a op m. (HasCallStack, Unbox a, Monoid a, Monoid op, Unbox op, PrimMonad m) => Int -> (Int -> a) -> m (LazySegmentTree a op (PrimState m))

\(O(N)\)

buildLSTree :: forall op a m. (HasCallStack, Unbox a, Monoid a, Monoid op, Unbox op, PrimMonad m) => Vector a -> m (LazySegmentTree a op (PrimState m))

\(O(N)\). TODO: Test it. Share the internal implementation with genearteLSTree takeing filling function.

sactLSTree :: forall a op m. (Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> Int -> op -> m ()

\(O(\log N)\) Appends the lazy operator monoid monoids over a span. They are just stored and propagated when queried.

sactAtLSTree :: forall a op m. (Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> op -> m ()

\(O(\log N)\) Acts on one leaf. TODO: Faster implementation.

foldLSTree :: forall a m op. (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> Int -> m a

\(O(\log N)\)

readLSTree :: forall a op m. (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> m a

\(O(\log N)\) Read one leaf. TODO: Faster implementation.

foldMayLSTree :: forall a op m. (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> Int -> m (Maybe a)

\(O(\log N)\)

foldAllLSTree :: forall a op m. (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> m a

\(O(\log N)\)

_propOpMonoidsToLeaf :: (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> m ()

\(O(\log N)\) Propagates the lazy operator monoids from top to bottom where the leaf vertex is contained.

iLeaf: Given with zero-based index.

_evalToRoot :: (HasCallStack, Monoid a, Unbox a, SemigroupAction op a, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> m ()

Evaluates parent values on updateSegmentTree. TODO: move to where clause of the update function?

bisectLSTree :: forall a m op. (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> Int -> (a -> Bool) -> m (Maybe Int, Maybe Int)

\(O(\log^2 N)\) The l, r indices are the zero-based leaf indices.

bisectLSTreeL :: forall a m op. (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> Int -> (a -> Bool) -> m (Maybe Int)

\(O(\log^2 N)\)

bisectLSTreeR :: forall a m op. (HasCallStack, Monoid a, Unbox a, Monoid op, SemigroupAction op a, Eq op, Unbox op, PrimMonad m) => LazySegmentTree a op (PrimState m) -> Int -> Int -> (a -> Bool) -> m (Maybe Int)

\(O(\log^2 N)\)