Data.SegmentTree.Strict

Description

Strict segment tree

Documentation

data SegmentTree v s a

A mutable segment tree backed by a complete binary tree.

Overview

A segment tree is a cache of a folding function. Each node corresponds to a folding range and the node contains the folding result.

A segment tree has a constant size and never be resized.

Operations

Modification takes \(O(log N)\), so creation takes \(N(log N)\). Lookup takes \(O(log N)\).

(Internal) Indices

The complete binary tree has 2 ^ depth - 1 elements.

Child elements of a parent node i has index 2 * i + 1 and 2 * i + 2.

The leaf indices start with length / 2 - 1.

Example:

           0
     1           2
  3     4     5     6
07 08 09 10 11 12 13 14

Constructors

SegmentTree (a -> a -> a) (v s a)

newSTreeG :: (MVector v a, PrimMonad m) => (a -> a -> a) -> Int -> a -> m (SegmentTree v (PrimState m) a)

Creates a segment tree for n leaves.

newSTreeV :: PrimMonad m => (a -> a -> a) -> Int -> a -> m (SegmentTree MVector (PrimState m) a)

Creates a segment tree of boxed elements.

newSTreeU :: (Unbox a, PrimMonad m) => (a -> a -> a) -> Int -> a -> m (SegmentTree MVector (PrimState m) a)

Creates a segment tree of unboxed elements.

newSTree :: (Unbox a, Monoid a, PrimMonad m) => Int -> m (SegmentTree MVector (PrimState m) a)

Creates a segment tree of unboxed monoids.

newSTree' :: (Unbox a, Semigroup a, PrimMonad m) => Int -> a -> m (SegmentTree MVector (PrimState m) a)

Creates a segment tree of unboxed monoid-likes. FIXME: Use NonNegative or Double' instead.

generateSTreeG :: (MVector v a, PrimMonad m) => (a -> a -> a) -> Int -> a -> (Int -> a) -> m (SegmentTree v (PrimState m) a)

Creates a segment tree.

generateSTreeV :: PrimMonad m => (a -> a -> a) -> Int -> a -> (Int -> a) -> m (SegmentTree MVector (PrimState m) a)

Creates a segment tree of boxed elements.

generateSTreeU :: (Unbox a, PrimMonad m) => (a -> a -> a) -> Int -> a -> (Int -> a) -> m (SegmentTree MVector (PrimState m) a)

Creates a segment tree of unboxed elements.

generateSTree :: (Unbox a, Monoid a, PrimMonad m) => Vector a -> m (SegmentTree MVector (PrimState m) a)

resetSTree :: (MVector v a, PrimMonad m) => SegmentTree v (PrimState m) a -> a -> m ()

Sets all the internal values of a segment tree to the given value which has to be zero.

REMARK: It takes lots of time. Consider a much more efficient resettiong strategy such as re-inserting zeros to used slots, or maybe use | compressInvNumG when you just need inversion number.

insertSTree :: (MVector v a, PrimMonad m) => SegmentTree v (PrimState m) a -> Int -> a -> m ()

Updates an SegmentTree leaf value and their parents up to top root.

modifySTree :: (MVector v a, PrimMonad m) => SegmentTree v (PrimState m) a -> (a -> a) -> Int -> m ()

Updates an SegmentTree leaf value and their parents up to top root.

_updateElement :: (HasCallStack, MVector v a, PrimMonad m) => SegmentTree v (PrimState m) a -> Int -> a -> m ()

(Internal) Updates an SegmentTree element (node or leaf) value and their parents up to top root. REMARK: It's faster to not INLINE the recursive function:

querySTree :: forall v a m. (HasCallStack, MVector v a, PrimMonad m) => SegmentTree v (PrimState m) a -> Int -> Int -> m (Maybe a)

Retrieves the folding result over the inclusive range `[l, r]` from SegmentTree.

unsafeFreezeSTree :: (Vector v a, PrimMonad m) => SegmentTree (Mutable v) (PrimState m) a -> m (v a)

Retrieves leaves.

WARNING: It doesn't remember correct number of leaves.