Data.Slide

Description

Sliding window data structures. See also SlideMin.

Typical Problems

Documentation

data StackSlidingFold s a

Stack-based sliding window folding, going from left to right only. SWAG (sliding window aggregation).

Constructors

StackSlidingFold

Fields

bufferSSF :: !(Buffer s a)

Stack of newly pushed elements, not yet moved to dualScanrSSF.

bufferFoldSSF :: !(MVector s a)

Fold of the elements in the buffer.

dualScanrSSF :: !(Buffer s a)

Right dual scanning, up to the last popped point. Works great with pop operations.

newSSF :: (PrimMonad m, Monoid a, Unbox a) => Int -> m (StackSlidingFold (PrimState m) a)

\(O(N)\)

clearSSF :: (PrimMonad m, Monoid a, Unbox a) => StackSlidingFold (PrimState m) a -> m ()

\(O(1)\)

pushBackSSF :: (PrimMonad m, Semigroup a, Unbox a) => StackSlidingFold (PrimState m) a -> a -> m ()

\(O(1)\)

popFrontSSF :: (PrimMonad m, Monoid a, Unbox a) => StackSlidingFold (PrimState m) a -> m ()

Amortised \(O(1)\).

foldSSF :: (PrimMonad m, Semigroup a, Unbox a) => StackSlidingFold (PrimState m) a -> m a

\(O(1)\)

data DequeSlidingFold s a

Dequeue-based sliding window folding. Prefer StackSlidingFold for speed. SWAG (sliding window aggregation).

Constructors

DequeSlidingFold

Fields

frontBufferDSF :: !(Buffer s a)

frontScanDSF :: !(Buffer s a)

backBufferDSF :: !(Buffer s a)

backScanDSF :: !(Buffer s a)

newDSF :: (PrimMonad m, Monoid a, Unbox a) => Int -> m (DequeSlidingFold (PrimState m) a)

\(O(N)\)

clearDSF :: (PrimMonad m, Monoid a, Unbox a) => DequeSlidingFold (PrimState m) a -> m ()

\(O(1)\)

pushFrontDSF :: (HasCallStack, PrimMonad m, Semigroup a, Unbox a) => DequeSlidingFold (PrimState m) a -> a -> m ()

\(O(1)\)

pushBackDSF :: (HasCallStack, PrimMonad m, Semigroup a, Unbox a) => DequeSlidingFold (PrimState m) a -> a -> m ()

\(O(1)\)

popFrontDSF :: (HasCallStack, PrimMonad m, Monoid a, Unbox a) => DequeSlidingFold (PrimState m) a -> m ()

Amortised \(O(1)\).

popBackDSF :: (HasCallStack, PrimMonad m, Monoid a, Unbox a) => DequeSlidingFold (PrimState m) a -> m ()

Amortised \(O(1)\).

foldDSF :: (HasCallStack, PrimMonad m, Semigroup a, Unbox a) => DequeSlidingFold (PrimState m) a -> m a

\(O(1)\)

balanceDSF :: (HasCallStack, PrimMonad m, Monoid a, Unbox a) => DequeSlidingFold (PrimState m) a -> m ()

\(O(N)\)