Data.SparseGraph

Description

vector-based sparse graph implementation (weightened or unweightened).

Heavily inspired by cojna/iota.

Documentation

type Edge = (Vertex, Vertex)

type WEdgeWith w = (Vertex, Vertex, w)

Weightened edge

type EdgeId = Int

data SparseGraph i w

CSR (compressed sparse row) representation of a graph (weightened or unweightened)

Constructors

SparseGraph

Fields

boundsSG :: !(i, i)

Vertex index boundary.

nVertsSG :: !Int

Number of vertices.

nEdgesSG :: !Int

Number of edges.

offsetsSG :: !(Vector Int)

Maps Vector to the starting edge index.

adjacentsSG :: !(Vector Vertex)

Adjacent vertices sorted with starting vertex.

edgeWeightsSG :: !(Vector w)

Edge weight information.

buildUSG :: Unindex i => (i, i) -> Vector (i, i) -> SparseGraph i ()

Builds an unweightned SparseGraph.

TODO: Faster implementation

buildWSG :: (Unindex i, Unbox w) => (i, i) -> Vector (i, i, w) -> SparseGraph i w

Builds a weightned SparseGraph.

buildRawSG :: (Unindex i, Unbox w) => (i, i) -> Vector (Vertex, Vertex, w) -> SparseGraph i w

adj :: SparseGraph i w -> Vertex -> Vector Vertex

Retrieves adjacent vertices.

eAdj :: SparseGraph i w -> Vertex -> Vector (EdgeId, Vertex)

Returns `(EdgeId, Vertex)` paris. Hardly used.

adjIx :: Unindex i => SparseGraph i w -> i -> Vector i

Retrieves adjacent vertex indices.

adjW :: Unbox w => SparseGraph i w -> Vertex -> Vector (Vertex, w)

Retrieves adjacent vertices with weights.

adjWIx :: (Unindex i, Unbox w) => SparseGraph i w -> i -> Vector (i, w)

Retrieves adjacent vertex indices with weights.

dfsSG :: Unindex i => SparseGraph i w -> i -> IxVector i (Vector Int)

componentsVecSG :: Ix i => SparseGraph i w -> i -> IxVector i (Vector Bool)

Also consider using union-find tree.

bfsSG :: Unindex i => SparseGraph i w -> i -> IxVector i (Vector Int)

djSG :: forall i w. (Unindex i, Num w, Ord w, Unbox w) => SparseGraph i w -> w -> i -> Vector w

Dijkstra: $O((E+V)log {V})$

dfsPathSG :: Unindex i => SparseGraph i w -> i -> i -> Maybe [Edge]

Returns a list of a route in reverse order (a route from end to start).

treeDfsPathSG :: Unindex i => SparseGraph i w -> i -> i -> [Edge]

Returns a list of a route in reverse order (a route from end to start).

topSortSG :: SparseGraph i w -> [Vertex]

topScc1SG :: forall i w m. PrimMonad m => SparseGraph i w -> MVector (PrimState m) Bool -> Vertex -> m [Vertex]

Partial running of scc over topologically sorted vertices, but for some connected components only.

revSG :: (Unindex i, Unbox w) => SparseGraph i w -> SparseGraph i w

Creates a reverse graph. TODO: return weightned graph

topSccSG :: (Unindex i, Unbox w) => SparseGraph i w -> [[Int]]

Collectes strongly connected components, topologically sorted. Upstream vertices come first, e.g., (v1 - v2) -> v3 -> v4.