Data.UnionFind.Mutable

Description

Union-find tree

Typical problems

Typical 012 - Red Painting (★4)

Documentation

newtype MUnionFind s

Dense, mutable union-find tree (originally by `@pel`)

Constructors

MUnionFind (MVector s MUFNode)

type IOUnionFind = MUnionFind RealWorld

type STUnionFind s = MUnionFind s

data MUFNode

`MUFChild parent | MUFRoot size`.

Constructors

MUFChild !Int

MUFRoot !Int

newMUF :: PrimMonad m => Int -> m (MUnionFind (PrimState m))

Creates a new Union-Find tree of the given size.

rootMUF :: PrimMonad m => MUnionFind (PrimState m) -> Int -> m Int

Returns the root node index.

groupsMUF :: PrimMonad m => MUnionFind (PrimState m) -> m IntSet

Returns all root vertices.

sameMUF :: PrimMonad m => MUnionFind (PrimState m) -> Int -> Int -> m Bool

Checks if the two nodes are under the same root.

_unwrapMUFRoot :: MUFNode -> Int

Just an internal helper.

uniteMUF :: PrimMonad m => MUnionFind (PrimState m) -> Int -> Int -> m ()

Unites two nodes.

sizeMUF :: PrimMonad m => MUnionFind (PrimState m) -> Int -> m Int

Returns the size of the root node, starting with `1`.