Data.UnionFind.Potencial

Description

Union-find tree under a differencial constraint system.

Typical problems

ABC 328 F - Good Set Query

Documentation

data PUnionFind s

Union-Find tree under a differencial constraint systems. Each vertex v is given potencial value p(v) which is a relative value tested on unifyPUF.

Implementation based on: https://qiita.com/drken/items/cce6fc5c579051e64fab

Constructors

DUnionFind

Fields

nodesPUF :: MVector s MUFNode

Node data (MUFParent size | MUFNode parent).

potencialPUF :: MVector s Int

Diffierencial potencial of each vertex.

newPUF :: PrimMonad m => Int -> m (PUnionFind (PrimState m))

Creates a new union-find tree under a differencial-potencal system.

rootPUF :: PrimMonad m => PUnionFind (PrimState m) -> Int -> m Int

Returns root to the given vertex, after updating the internal differencial potencials.

unifyPUF :: PrimMonad m => PUnionFind (PrimState m) -> Int -> Int -> Int -> m Bool

Unifies two nodes, managing their differencial potencial so that p(v1) - p(v2) = d holds. Returns True when thry're newly unified. Returns False if they're already unified or when the unification didn't match the existing potencials.

sizePUF :: PrimMonad m => PUnionFind (PrimState m) -> Int -> m Int

Returns the number of vertices belonging to the same group.

samePUF :: PrimMonad m => PUnionFind (PrimState m) -> Int -> Int -> m Bool

Has the same root / belongs to the same group.

canUnifyPUF :: PrimMonad m => PUnionFind (PrimState m) -> Int -> Int -> Int -> m Bool

Can be unified, keeping the equiation p(v1) - p(v2) = d.

potPUF :: PrimMonad m => PUnionFind (PrimState m) -> Int -> m Int

Returns p(v): the potencial of the vertex in their group.

diffPUF :: PrimMonad m => PUnionFind (PrimState m) -> Int -> Int -> m Int

Returns p(v1) - p(v2).

clearPUF :: PrimMonad m => PUnionFind (PrimState m) -> m ()