ToyLib.DP

Description

Typical DP utilities

TODO: Refactor relaxMany variants.

Documentation

constructFor :: (Unbox a, Unbox b) => a -> Vector b -> (Vector a -> b -> a) -> Vector a

Variant of constructN.

relaxMany :: (HasCallStack, Vector v a, Vector v (Int, a), Vector v b) => (a -> a -> a) -> v a -> v b -> (b -> v (Int, a)) -> v a

accumulate variant with concatMap-like expander. Be wanrned that *the input is consumed in-place*. Run like relaxMany vec0 (U.force vec0) $ x -> .. if it needs to be cloned.

relaxMany !f !vec0 !input !expander ~ G.accumulate f vec0 $ G.concatMap expander input

irelaxMany :: (HasCallStack, Vector v a, Vector v (Int, a), Vector v b) => (a -> a -> a) -> v a -> v b -> (Int -> b -> v (Int, a)) -> v a

relaxMany with index input. Be wanrned that *the input is consumed in-place*. Run like relaxMany vec0 (U.force vec0) $ x -> .. if it needs to be cloned.

relaxMany' :: (Monoid m, Unbox m, Unbox a) => Vector m -> Vector a -> (a -> Vector (Int, m)) -> Vector m

Monoid variant of relaxMany

pushBasedConstructN :: (HasCallStack, Vector v a, Vector v (Int, a)) => (a -> a -> a) -> v a -> (Int -> v a -> v (Int, a)) -> v a

spansU :: Int -> Int -> Vector ((Int, Int), (Int, Int))

Returns non-zero two spans over the given inclusive range [l, r]. >>> spansU 3 6 [((3,3),(4,6)),((3,4),(5,6)),((3,5),(6,6))]

constructIV :: (Unindex i, Unbox a) => (i, i) -> (IxVector i (Vector a) -> i -> a) -> IxVector i (Vector a)

constructN for IxVector

spanDP :: Unbox a => Int -> a -> (Int -> a) -> (IxVector (Int, Int) (Vector a) -> (Int, Int) -> a) -> IxVector (Int, Int) (Vector a)

Span-based DP with preset index patterns.

tspDP :: Int -> IxUVector (Int, Int) Int -> Vector Int

Typical set-based DP.

Typical problems

ABC 317 C - Remembering the Days

enumerateBitSets :: Int -> [[Int]]

Enumerates all possible bitsets that composes bit n - 1.

>>> enumerateBitSets 4
[[8,4,2,1],[12,2,1],[8,6,1],[4,10,1],[14,1],[8,4,3],[12,3],[8,2,5],[10,5],[8,7],[4,2,9],[6,9],[4,11],[2,13],[15],[]]

Typical problems

ABC 310 D - Peaceful Teams

ABC 319 D - General Weighted Max Matching (Not the exact pattern though)

lcs :: HasCallStack => Vector Int -> Int

The input must be one-based