ToyLib.IO

Description

Super dirty slow IO

Main procedures

ints2, intsW, parse, digitsU, getGrid and charsH.

ReadBS and parse

Primitives are ReadBS. Tuples of ReadBS are also ReadBS:

>>> convertBS @(Int, Char, String, Float) $ BS.pack "42 c string 2.5"
(42,'c',"string",2.5)

Vectors are ReadBS and they can also be embedded in the end of a tuple:

>>> convertBS $ BS.pack "1 string 3.5 10 20 30 40" :: (Int, String, Float, U.Vector Int)
(1,"string",3.5,[10,20,30,40])

Documentation

int :: IO Int

ints :: IO [Int]

Reads one line as a list of integers.

class ReadBS a where

Read from a space-delimited ByteStrtig.

Minimal complete definition

Nothing

Methods

convertBS :: ByteString -> a

default convertBS :: Read a => ByteString -> a

readBS :: ByteString -> (a, ByteString)

For use with unfoldrExactN.

readMayBS :: ByteString -> Maybe (a, ByteString)

For use with unfoldr.

convertG :: (ReadBS a, Vector v a) => ByteString -> v a

Converrts the given ByteString as a vector of a.

convertNG :: (ReadBS a, Vector v a) => Int -> ByteString -> v a

Converts the given ByteString as a vector of a with n elements.

parse :: ReadBS a => IO a

Parses one line via the ReadBS class.

ints1 :: IO Int

ints2 :: IO (Int, Int)

ints3 :: IO (Int, Int, Int)

ints4 :: IO (Int, Int, Int, Int)

ints5 :: IO (Int, Int, Int, Int, Int)

ints6 :: IO (Int, Int, Int, Int, Int, Int)

intsW :: Vector v Int => Int -> IO (v Int)

intsG :: Vector v Int => IO (v Int)

Reads one line as an integer.

intsV :: IO (Vector Int)

intsU :: IO (Vector Int)

Reads one line as a vector of integers.

digitsU :: IO (Vector Int)

intsRestG :: Vector v Int => IO (v Int)

intsRestU :: IO (Vector Int)

getGraph :: Int -> Int -> IO (Array Int [Int])

Creates a graph from 1-based vertices

convertNBS :: forall a. (Unbox a, ReadBS a) => Int -> Vector ByteString -> Vector a

Converts n lines of **whitespace-delimited ByteString** into a flat vector of type a.

>>> convertNBS @Int (3 * 3) $ V.map BS.pack $ V.fromList ["1 2 3", "4 5 6", "7 8 9"]
[1,2,3,4,5,6,7,8,9]

getHW :: (Unbox a, ReadBS a) => Int -> Int -> IO (Vector a)

Reads h lines of stdin and converts them as HxW **whitespace-delimited ByteString** and converts them into a flat vector of type a.

getGrid :: Int -> Int -> IO (IxVector (Int, Int) (Vector Int))

Reads h lines of stdin and converts them into a IxVector reading as HxW **whitespace-separated** input.

convertCharsHW :: Vector ByteString -> Vector Char

Converts @n` lines of ByteString into a flat vector.

>>> U.map (== '#') . convertCharsHW $ V.map BS.pack $ V.fromList ["#.#", ".#."]
[True,False,True,False,True,False]

charsH :: Int -> IO (Vector Char)

See convertCharsHW.

concat2 :: [(a, a)] -> [a]

concat two-item tuples

concatMap2 :: (a -> (b, b)) -> [a] -> [b]

getWGraph :: Int -> Int -> IO (Array Int [Entry Int Int])

Creates a weightend graph from 1-based vertices

getWGraph0 :: Int -> Int -> IO (Array Int [Entry Int Int])

Creates a weightend graph from 1-based vertices

endlBSB :: Builder

putBSB :: Builder -> IO ()

putLnBSB :: Builder -> IO ()

class ShowBSB a where

Show as a bytestring builder.

Minimal complete definition

Nothing

Methods

showBSB :: a -> Builder

default showBSB :: Show a => a -> Builder

showLnBSB :: ShowBSB a => a -> Builder

printBSB :: ShowBSB a => a -> IO ()

concatBSB :: Vector v a => (a -> Builder) -> v a -> Builder

See unwordsBSB as example.

unwordsBSB :: (ShowBSB a, Vector v a) => v a -> Builder

unlinesBSB :: (ShowBSB a, Vector v a) => v a -> Builder

yn :: Bool -> String

ynBSB :: Bool -> Builder

printYn :: Bool -> IO ()