ToyLib.Parser

Description

StateT -based parser.

Documentation

runIO :: StateT ByteString IO a -> IO a

Reads the whole stdin and runs the user program.

runFileIO :: StateT ByteString IO a -> String -> IO a

Reads a file runs the user program.

Primitives

int' :: MonadState ByteString m => m Int

Parses an Int.

int1' :: MonadState ByteString m => m Int

Parses an Int and substracts one.

char' :: MonadState ByteString m => m Char

word' :: MonadState ByteString m => m ByteString

double' :: MonadState ByteString m => m Double

Parsers an Int and converts it into a Double.

Tuples

ints2' :: MonadState ByteString m => m (Int, Int)

ints11' :: MonadState ByteString m => m (Int, Int)

ints3' :: MonadState ByteString m => m (Int, Int, Int)

ints110' :: MonadState ByteString m => m (Int, Int, Int)

ints011' :: MonadState ByteString m => m (Int, Int, Int)

ints111' :: MonadState ByteString m => m (Int, Int, Int)

ints4' :: MonadState ByteString m => m (Int, Int, Int, Int)

ints5' :: MonadState ByteString m => m (Int, Int, Int, Int, Int)

ints6' :: MonadState ByteString m => m (Int, Int, Int, Int, Int, Int)

Readers

line' :: MonadState ByteString m => m ByteString

Reads one line from the state. FIXME: It ignores empty lines.

withLine' :: MonadState ByteString m => State ByteString a -> m a

Reads one line from the state and runs a pure parser for it.

More

ints' :: MonadState ByteString m => m [Int]

Reads one line an unboxed vector.

intsU' :: MonadState ByteString m => m (Vector Int)

Reads one line an unboxed vector.

intsN' :: MonadState ByteString m => Int -> m (Vector Int)

Reads n values as an unboxed vector.

digitsU' :: MonadState ByteString m => m (Vector Int)

Reads one line as digits. TODO: one word might be better